% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.matrix.xtabs.R
\name{as.array.xtabs}
\alias{as.array.xtabs}
\alias{as.matrix.xtabs}
\title{Coerce \code{xtabs} Object to Array or Matrix}
\usage{
\method{as.array}{xtabs}(x, ...)

\method{as.matrix}{xtabs}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"xtabs"}.}

\item{...}{Ignored.}
}
\value{
An array (or matrix in the two-dimensional case) with no
additional classes.
}
\description{
Converts an object of class \code{"xtabs"} to a plain array or matrix by
dropping all additional classes such as \code{"xtabs"} and \code{"table"},
along with the \code{call} attribute.
}
\examples{
xt <- xtabs(~ cyl + gear, data = mtcars)
class(as.matrix(xt))
# "matrix" "array"

}
\seealso{
Other data.manipulation:
\code{\link[=appendEnum]{appendEnum()}},
\code{\link[=appendRowNames]{appendRowNames()}},
\code{\link[=appendX]{appendX()}},
\code{\link[=asBinary]{asBinary()}},
\code{\link[=columnWrap]{columnWrap()}},
\code{\link[=combLevels]{combLevels()}},
\code{\link[=dummy]{dummy()}},
\code{\link[=nf]{nf()}},
\code{\link[=recodeX]{recodeX()}},
\code{\link[=recycle]{recycle()}},
\code{\link[=renameX]{renameX()}},
\code{\link[=revCode]{revCode()}},
\code{\link[=revX]{revX()}},
\code{\link{setAttr-removeAttr-keepAttr}},
\code{\link[=setNamesX]{setNamesX()}},
\code{\link[=sortX]{sortX()}},
\code{\link[=splitAt]{splitAt()}},
\code{\link[=splitX]{splitX()}},
\code{\link[=stringsAsFactors]{stringsAsFactors()}},
\code{\link[=toBaseR]{toBaseR()}}
}
\concept{data.manipulation}
\concept{table}
