% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charAsc.R
\name{charToAsc}
\alias{charToAsc}
\alias{ascToChar}
\title{Convert ASCII Codes to Characters and Vice Versa}
\usage{
charToAsc(x)

ascToChar(i)
}
\arguments{
\item{x}{vector of strings.}

\item{i}{numeric (integer) vector of values in \code{1:255}.}
}
\value{
\code{AscToChar} returns a vector of the same length as i.
\code{CharToAsc} returns a list of numeric vectors of character length of
each string in x.
}
\description{
\code{ascToChar()} returns a character for each ASCII code (integer) supplied.\cr
\code{charToAsc()} returns integer codes in \code{0:255} for each (one byte)
character in all strings in \code{x}.
}
\details{
Only codes in \code{1:127} make up the ASCII encoding which should be
identical for all \R versions, whereas the \emph{\sQuote{upper}} half is
often determined from the ISO-8859-1 (aka \dQuote{ISO-Latin 1)} encoding,
but may well differ, depending on the locale setting, see also
\code{\link{Sys.setlocale}()}.

Note that \code{0} is no longer allowed since, \R does not allow \code{\0}
aka \code{nul} characters in a string anymore.
}
\examples{

(x <- charToAsc("Silvia"))

# will be pasted together
ascToChar(x)

# use strsplit if the single characters are needed
strsplit(ascToChar(x), split=NULL)

# this would be an alternative, but the latter would be of class raw
decToHex(charToAsc("Silvia"))
# same as:
charToRaw("Silvia")

}
\seealso{
\code{\link{charToRaw}}

Other string.utilities: 
\code{\link{strSplitToCol}()},
\code{\link{strSplitToDummy}()}
}
\author{
unknown guy out there, help text partly taken from M. Maechler's
\pkg{sfsmisc}.
}
\concept{data-manipulation}
\concept{string-manipulation}
\concept{string.utilities}
\keyword{manip}
