% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combSet.R
\name{combSet}
\alias{combSet}
\title{Samples for Combinations of a Set}
\usage{
combSet(x, m, repl = FALSE, ord = FALSE, as.list = FALSE)
}
\arguments{
\item{x}{a vector of numeric values or characters. Characters need not be
unique.}

\item{m}{number of elements to choose. For \code{combSet} can \code{m} be a
numeric vector too.}

\item{repl}{logical. Should repetition of the same element be allowed?
Defaults to FALSE}

\item{ord}{logical. Does the order matter? Default is FALSE.}

\item{as.list}{logical, defining if the results should be returned in a flat
list, say every sample is a single element of the resulting list. Default is
FALSE.}
}
\value{
a matrix with all possible combinations of the
values in x\cr if m contains more than
one element the result will be a list of matrices or a flat list if
\code{as.list} is set to \code{TRUE}
}
\description{
Return the value sets of combinations.
}
\examples{
x <- letters[1:4]
m <- 2

# the samples
combSet(x, m, repl=TRUE, ord=FALSE)
combSet(x, m, repl=TRUE, ord=TRUE)
combSet(x, m, repl=FALSE, ord=TRUE)
combSet(x, m, repl=FALSE, ord=FALSE)

# build all subsets of length 1, 3 and 5 and return a flat list
x <- letters[1:5]
combSet(x=x, m=c(1, 3, 5), as.list=TRUE)

}
\seealso{
\code{\link{combPairs}},
\code{\link{combn}}, \code{\link{choose}}, \code{\link{factorial}},
\cr \code{vignette("Combinatorics")}

Other combinatorics: 
\code{\link{combN}()},
\code{\link{combPairs}()},
\code{\link{permn}()},
\code{\link{rndGroupSplit}()},
\code{\link{sampleX}()}
}
\concept{combinatorics}
\concept{mathematics}
\concept{vector-manipulation}
