% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combPairs.R
\name{combPairs}
\alias{combPairs}
\title{Get All Pairs Out of One or Two Sets of Elements}
\usage{
combPairs(x, y = NULL)
}
\arguments{
\item{x}{a vector of elements, must contain at least 2 elements if
\code{y} is \code{NULL}.}

\item{y}{a vector of elements, need not be same dimension as x.  If y is not
\code{NULL} then all combination x and y are returned.}
}
\value{
a data.frame with two columns \code{X1} and \code{X2}
containing the pairwise combinations.
}
\description{
Returns all combinations of 2 out of the elements in x or x and y (if
defined). Combinations of the same elements will be dropped (no replacing).
The vector \code{x} need not contain unique values. Duplicate elements
in \code{x} will result in duplicate pairs.
}
\details{
If y = \code{NULL} then all combination of 2 out of x are returned. \cr If y
is defined then all combinations of x and y are calculated.
}
\examples{

combPairs(letters[1:4])
combPairs(x = letters[1:4], y = LETTERS[1:2])

# get all pairs of combinations between factors and numerics out of a data.frame
combPairs(which(sapply(CO2, is.numeric)), which(sapply(CO2, is.factor)))

}
\seealso{
\code{\link[=combn]{combn()}}, \code{\link[=expand.grid]{expand.grid()}}, \code{\link[=outer]{outer()}}, \code{\link[=lower.tri]{lower.tri()}}

Other combinatorics:
\code{\link[=combN]{combN()}},
\code{\link[=combSet]{combSet()}},
\code{\link[=pairApply]{pairApply()}},
\code{\link[=permn]{permn()}},
\code{\link[=randGroupSplit]{randGroupSplit()}},
\code{\link[=sampleX]{sampleX()}}
}
\concept{combinatorics}
\concept{number-theory}
