% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{distance}
\alias{distance}
\title{Interval distance}
\usage{
distance(x, y)
}
\arguments{
\item{x}{Numeric vector of length 2 or matrix with 2 columns}

\item{y}{Numeric vector of length 2 or matrix with 2 columns}
}
\value{
Numeric vector of distances (0 if overlapping)
}
\description{
Computes the distance between non-overlapping intervals.
Returns 0 if intervals overlap (including boundary touching).
}
\examples{
distance(c(1, 2), c(4, 5))  # 2
distance(c(1, 5), c(3, 7))  # 0
distance(c(1, 3), c(3, 5))  # 0

}
