% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractArgs.R
\name{extractArgs}
\alias{extractArgs}
\title{Extract named arguments from dots with defaults}
\usage{
extractArgs(dots, defaults, validate = NULL, returnRest = FALSE)
}
\arguments{
\item{dots}{Named list of arguments (usually \code{list(...)})}

\item{defaults}{Named list of default values}

\item{validate}{Optional validation function}

\item{returnRest}{Logical; return unused arguments}
}
\value{
Named list of extracted arguments (and optionally remaining ones)
}
\description{
Utility to extract a subset of arguments from a list (typically \code{list(...)})
and merge them with default values.
}
\seealso{
Other pkg.introspection:
\code{\link[=funArgs]{funArgs()}},
\code{\link[=funCalls]{funCalls()}},
\code{\link[=funKeywords]{funKeywords()}},
\code{\link[=funList]{funList()}},
\code{\link[=getDotsArg]{getDotsArg()}},
\code{\link[=mergeArgs]{mergeArgs()}},
\code{\link[=quot]{quot()}},
\code{\link[=rdLabels]{rdLabels()}},
\code{\link[=rdTitle]{rdTitle()}},
\code{\link[=strX]{strX()}}
}
\concept{pkg.introspection}
\concept{programming}
