% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fibonacci.R
\name{fibonacci}
\alias{fibonacci}
\title{Fibonacci Numbers}
\usage{
fibonacci(n)
}
\arguments{
\item{n}{nonnegative integer (<= 78) or vector of such integers}
}
\value{
An integer-valued numeric vector.
}
\description{
Generate Fibonacci numbers. The Fibonacci numbers can also be calculated
using the golden ratio \code{phi}, as demonstrated in the examples.
}
\details{
Generates the \code{n}-th Fibonacci number, whereas \code{fibonacci(0) = 0}.
\cr The golden ratio is defined as \code{phi = 0.5*(1+sqrt(5))}.

Values of \code{n} are limited to 78, as larger Fibonacci numbers exceed
the range in which doubles represent integers exactly (\code{2^53}).
}
\examples{

fibonacci(0)                            # 0
fibonacci(2)                            # 1
fibonacci(0:3)                          # 0 1 1 2
fibonacci(0:25)                         # ... 75025 121393

# Golden ratio = Fib(25)/ Fib(24)
f25 <- quot(fibonacci(24:25))           # 1.618033989
phi <- (sqrt(5) + 1)/2
abs(f25 - phi)                          # 7.945178e-11

# Fibonacci numbers without iteration
fibo <- function(n) {
  phi <- (sqrt(5) + 1)/2
  fib <- (phi^(n+1) - (1-phi)^(n+1)) / (2*phi - 1)
  round(fib)
}

fibo(30:33)                             # 1346269 2178309 3524578 5702887

}
\references{
\url{https://en.wikipedia.org/wiki/Fibonacci_number}\cr
\url{https://mathworld.wolfram.com/GoldenRatio.html}
}
\seealso{
Other number.theory:
\code{\link{GCD-LCM}},
\code{\link[=digitSum]{digitSum()}},
\code{\link[=divisors]{divisors()}},
\code{\link[=factorize]{factorize()}},
\code{\link[=isOdd]{isOdd()}},
\code{\link[=isPrime]{isPrime()}},
\code{\link[=primes]{primes()}}
}
\concept{number-theory}
\concept{number.theory}
\concept{numerical-methods}
