% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funCalls.R
\name{funCalls}
\alias{funCalls}
\title{List Calls Used in Function}
\usage{
funCalls(fun, package = NULL, sorted = FALSE)
}
\arguments{
\item{fun}{the name of the function.}

\item{package}{optional name of a package; if given, the result is
filtered to source environments matching \code{package}.}

\item{sorted}{logical; whether calls are sorted alphabetically. Defaults to
\code{FALSE}.}
}
\value{
a list of character vectors with the function calls, grouped by
the environment the called functions were found in.
}
\description{
For screening purposes it can be useful to get a list of all function calls
our function may depend on. \code{funCalls()} parses the function
source and returns all found function calls grouped by their package.
}
\details{
The source packages are resolved via \code{\link[=find]{find()}}, which only
sees attached packages. Calls to functions from packages that are not on
the search path are reported under \code{"<not found>"}.
}
\note{
Based on code by Nicholas Cooper, adapted to conform to package
standards.
}
\examples{

funCalls("combN", package="bedrock")

}
\seealso{
\code{\link[=getParseData]{getParseData()}}

Other pkg.funinfo:
\code{\link[=auditNames]{auditNames()}},
\code{\link[=funArgs]{funArgs()}},
\code{\link[=funKeywords]{funKeywords()}},
\code{\link[=funList]{funList()}},
\code{\link[=rdLabels]{rdLabels()}},
\code{\link[=rdTitle]{rdTitle()}}
}
\concept{introspection}
\concept{pkg.funinfo}
\concept{programming}
