% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funList.R
\name{funList}
\alias{funList}
\title{List Functions in a Package}
\usage{
funList(package, exported = TRUE)
}
\arguments{
\item{package}{the name of the package.}

\item{exported}{logical; whether only exported functions are listed.
Defaults to \code{TRUE}.}
}
\value{
a sorted character vector with the function names.
}
\description{
List all the functions in a package.
}
\details{
This is just a wrapper for the namespace inspection functions (as I always
forgot how to do the trick). By default only the exported functions are
returned; with \code{exported = FALSE} all functions defined in the
package namespace are listed, including internal ones.
}
\examples{

funList("bedrock")

}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) \emph{The
New S Language}. Wadsworth & Brooks/Cole.
}
\seealso{
\code{\link[=ls]{ls()}}, \code{\link[=ls.str]{ls.str()}}, \code{\link[=lsf.str]{lsf.str()}},
\code{\link[=getNamespaceExports]{getNamespaceExports()}}

Other pkg.funinfo:
\code{\link[=auditNames]{auditNames()}},
\code{\link[=funArgs]{funArgs()}},
\code{\link[=funCalls]{funCalls()}},
\code{\link[=funKeywords]{funKeywords()}},
\code{\link[=rdLabels]{rdLabels()}},
\code{\link[=rdTitle]{rdTitle()}}
}
\concept{introspection}
\concept{pkg.funinfo}
\concept{programming}
