% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDotsArg.R
\name{getDotsArg}
\alias{getDotsArg}
\title{Get a Single Argument from Dots with Default}
\usage{
getDotsArg(dots, name, default = NULL)
}
\arguments{
\item{dots}{named list (usually \code{list(...)}).}

\item{name}{character string, argument name.}

\item{default}{default value if argument not present.}
}
\value{
the value of the argument or \code{default}.
}
\description{
Lightweight helper to extract a single named argument from a list (typically
\code{list(...)}). If the argument is not present, a default value is
returned.
}
\examples{
f <- function(...) {
  dots <- list(...)
  getDotsArg(dots, "col", default = "black")
}

f(col = "red", lwd = 2)
f(lwd = 2)

}
\seealso{
For extracting several arguments
at once use \code{\link[=extractArgs]{extractArgs()}}.

Other pkg.args:
\code{\link[=callIf]{callIf()}},
\code{\link[=extractArgs]{extractArgs()}},
\code{\link[=mergeArgs]{mergeArgs()}},
\code{\link[=recycle]{recycle()}}
}
\concept{introspection}
\concept{pkg.args}
\concept{programming}
