% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range-operators.R
\name{\%:\%}
\alias{\%:\%}
\title{Select a Range Between Two Elements (Lazy)}
\usage{
x \%:\% rng
}
\arguments{
\item{x}{a vector.}

\item{rng}{a vector of length 2: \code{c(from, to)}. May contain \code{NA} to match
missing values in \code{x}.}
}
\value{
a subset of \code{x}, from the first occurrence of \code{rng[1]} to the
first occurrence of \code{rng[2]}.
}
\description{
Returns the elements of \code{x} from the first occurrence of \code{rng[1]} to the
first occurrence of \code{rng[2]}. Equivalent to a non-greedy ("lazy") match,
analogous to lazy quantifiers in regular expressions.
}
\examples{
letters \%:\% c("c", "g")

x <- c("a", "b", "c", "d", "c", "e", "f", "c")
x \%:\% c("c", "e")

colnames(Pizza) \%:\% c("price", "nps")

}
\seealso{
the \code{\link{\%::\%}} operator for the greedy variant, which uses the
\emph{last} occurrence of \code{rng[2]}.
}
