% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nin.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Find Matching (or Non-Matching) Elements}
\usage{
x \%nin\% table
}
\arguments{
\item{x}{a vector (numeric, character, factor)}

\item{table}{a vector (numeric, character, factor), matching the mode of
\code{x}}
}
\value{
vector of logical values with length equal to length of \code{x}.
}
\description{
\code{\%nin\%} is a binary operator, which returns a logical vector indicating
if there is a match or not for its left operand. A true vector element
indicates no match in left operand, false indicates a match.
}
\examples{

c('a','b','c') \%nin\% c('a','b')

}
\seealso{
\code{\link{match}}, \code{\link{\%in\%}}, \code{\link{between}}
}
\author{
Frank E Harrell Jr \href{mailto:f.harrell@vanderbilt.edu}{f.harrell@vanderbilt.edu} \%\% ~~who you are~~
}
\keyword{character}
\keyword{manip}
