% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{\%overlaps\%}
\alias{\%overlaps\%}
\title{Interval overlap operator}
\usage{
x \%overlaps\% y
}
\arguments{
\item{x, y}{Numeric intervals}
}
\value{
Logical vector
}
\description{
Convenience operator for interval overlap.
}
\examples{
c(1,5) \%overlaps\% c(3,7)

}
