% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inDots.R
\name{inDots}
\alias{inDots}
\title{Extract Argument from \code{...} with Default}
\usage{
inDots(..., arg, default = NULL)
}
\arguments{
\item{...}{Arguments passed to the calling function.}

\item{arg}{A character string specifying the name of the argument to extract.}

\item{default}{Value returned if \code{arg} is not found.}
}
\value{
The value of the argument \code{arg} if present, otherwise \code{default}.
}
\description{
Retrieves a named argument from \code{...}. If the argument is not present,
a default value is returned. If the argument appears multiple times,
the last occurrence is used (consistent with base R behavior).
}
\details{
Only named arguments in \code{...} are considered. Unnamed arguments are ignored.
}
\examples{
f <- function(...) {
  inDots(..., arg = "col", default = "black")
}

f(col = "red")
f(lwd = 2)
f(col = "blue", col = "green")  # returns "green"

}
\seealso{
Other pkg.introspection:
\code{\link[=funArgs]{funArgs()}},
\code{\link[=funCalls]{funCalls()}},
\code{\link[=funKeywords]{funKeywords()}},
\code{\link[=funList]{funList()}},
\code{\link[=getRdLabels]{getRdLabels()}},
\code{\link[=mergeArgs]{mergeArgs()}},
\code{\link[=strX]{strX()}}
}
\concept{data-manipulation}
\concept{package-utilities}
\concept{pkg.introspection}
