% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n.R
\name{n}
\alias{n}
\title{Alias for \code{as.numeric}}
\usage{
n()
}
\value{
A numeric vector.
}
\description{
A short alias for \code{\link{as.numeric}}.
}
\details{
This is a convenience alias for \code{as.numeric}, useful in interactive
data analysis for more concise code.
}
\examples{
n("123")
n(c("1", "2", "3"))

}
\seealso{
\code{\link{as.numeric}}
}
