% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naIf.R
\name{naIf}
\alias{naIf}
\title{Replace Values with NA}
\usage{
naIf(x, values)
}
\arguments{
\item{x}{a vector.}

\item{values}{values to be replaced by \code{NA}.}
}
\value{
a vector of the same type as \code{x}.
}
\description{
Replaces specified values in a vector with \code{NA}, in the manner of
SQL's \code{NULLIF}. This is the complementary operation to
\code{\link[=coalesceX]{coalesceX()}}.
}
\examples{
naIf(c(1, 2, 99, 3, 99), 99)
naIf(c("a", "b", "n/a", ""), c("n/a", ""))

}
\seealso{
Other vector.na:
\code{\link[=coalesceX]{coalesceX()}},
\code{\link[=isNA]{isNA()}},
\code{\link[=locf]{locf()}},
\code{\link[=naReplace]{naReplace()}}
}
\concept{data-inspection}
\concept{missing-value}
\concept{vector.na}
