% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz.R
\name{nz}
\alias{nz}
\title{Extract Non-Zero Values}
\usage{
nz(x, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{a numeric vector.}

\item{tol}{tolerance passed to \code{\link[=isZero]{isZero()}}.}
}
\value{
a vector containing only the non-zero elements of \code{x}.
}
\description{
Returns all non-zero elements of a vector. Zeroness is determined by
\code{\link[=isZero]{isZero()}}, i.e. within a numerical tolerance.
}
\details{
\code{NA} elements are not considered zero and are retained in the
result.
}
\examples{
nz(c(0, 1, 2, 0, 3))
nz(c(1e-20, 1, NA))

}
\seealso{
\code{\link[=isZero]{isZero()}}

Other vector.utils:
\code{\link[=unwhich]{unwhich()}}
}
\concept{data-inspection}
\concept{vector.utils}
