% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdfManual.R
\name{pdfManual}
\alias{pdfManual}
\title{Open CRAN PDF Manual of a Package}
\usage{
pdfManual(package)
}
\arguments{
\item{package}{package name (symbol or character).}
}
\value{
the URL of the PDF manual, invisibly. Called for its side effect
of opening the browser.
}
\description{
PDF versions of the manual are usually not included as vignettes in R
packages. Still this format is convenient for reading and doing full text
search. \cr This function creates the appropriate link to the pdf file on
CRAN and opens the pdf manual in a browser window.
}
\details{
A warning (not an error) is issued if the package is not installed
locally, as the manual may well exist on CRAN anyway.
}
\examples{
# opens a browser window, hence only run in an interactive session
if (interactive()) {
  pdfManual(DescToolsX)
  pdfManual("bedrock")
}

}
\seealso{
\code{\link[=browseURL]{browseURL()}}

Other file.io:
\code{\link[=parseSasDatalines]{parseSasDatalines()}},
\code{\link[=peekFile]{peekFile()}},
\code{\link[=readDownload]{readDownload()}}
}
\concept{file.io}
\concept{introspection}
