% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range-operators.R
\name{range-operators}
\alias{range-operators}
\alias{\%:\%}
\alias{\%::\%}
\title{Select a Range Between Two Elements}
\usage{
x \%:\% rng

x \%::\% rng
}
\arguments{
\item{x}{a vector.}

\item{rng}{a vector of length 2: \code{c(from, to)}. May contain \code{NA} to match
missing values in \code{x}.}
}
\value{
a subset of \code{x}, from the first occurrence of \code{rng[1]} to the
first (\verb{\%:\%}) or last (\verb{\%::\%}) occurrence of \code{rng[2]}.
}
\description{
Returns the elements of \code{x} from the first occurrence of \code{rng[1]} up to an
occurrence of \code{rng[2]}. The two operators differ in which occurrence of
the end value terminates the range, analogous to lazy and greedy
quantifiers in regular expressions:
}
\details{
\itemize{
\item \verb{\%:\%} (lazy): up to the \emph{first} occurrence of \code{rng[2]}.
\item \verb{\%::\%} (greedy): up to the \emph{last} occurrence of \code{rng[2]}.
}
}
\examples{
letters \%:\% c("c", "g")

x <- c("a", "b", "c", "d", "c", "e", "f", "c")
x \%:\%  c("c", "e")
x \%::\% c("c", "c")     # greedy: up to the last "c"

# select a column range by name
colnames(mtcars) \%:\% c("hp", "vs")

}
\seealso{
Other data.interval:
\code{\link{between-operators}},
\code{\link{intervals}}
}
\concept{data.interval}
\concept{range}
