% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeX.R
\name{recodeX}
\alias{recodeX}
\title{Recode a Variable}
\usage{
recodeX(
  x,
  ...,
  keep = NULL,
  elseLevel = NA,
  ref = NULL,
  useEmpty = FALSE,
  num = FALSE
)
}
\arguments{
\item{x}{the factor whose levels are to be altered.  If x is
\code{character} it will be factorized (using \code{factor} defaults) but
returned as \code{character} again.}

\item{\dots}{the old levels (combined by \code{c}() if there are several)
named with the new level:\cr \verb{newlevel_a = c("old_a", "old_b"), }\cr\code{newlevel_b = c("old_c", "old_d")}\cr See examples.}

\item{keep}{vector of levels that should be left untouched.}

\item{elseLevel}{the value for levels, which are not matched by newlevel
list.  If this is set to \code{NULL}, the elseLevels will be left unchanged.
If set to \code{NA} (default) non matched levels will be set to \code{NA}.}

\item{ref}{the reference level, typically a string.}

\item{useEmpty}{logical. Defines how a new level, which can't be found in
x, should be handled.  Should it be left in the level's list or be dropped?
The default is \code{FALSE}, which drops empty levels.}

\item{num}{logical. If set to \code{TRUE} the result will be numeric. This
is useful if you want to recode strings to specific numeric values.}
}
\value{
the factor having the new levels applied.\cr if \code{x} was a
\code{character} vector, the result will also be \code{character}.
}
\description{
Combining or rearranging a factor can be tedious if it has many levels.
\code{recodeX()} supports this step by accepting a direct definition of new
levels by enumerating old levelnames as argument and adding an
\code{"elseLevel"} option. If new levels are given as integer values they
will be translated in the according levels.
}
\examples{

set.seed(1984)
x <- factor(sample(1:15, 20, replace=TRUE))
levels(x) <- paste("old", levels(x), sep="_")

y <- recodeX(x,
            "new_1"   = c("old_1","old_4","old_5"),
            "new_2"   = c("old_6","old_10","old_11"),
            "new_3"   = c("old_12","old_13"),
            elseLevel = "other")
data.frame(x=x, y=y)

# Coding NAs, NA is recoded to new_1
x[5:6] <- NA
x <- x[1:7] 

data.frame(
  x, 
  RecodeNA = recodeX(x,
                    "new_1"   = c("old_4","old_8", NA),
                    elseLevel = "other"),
       
  # NAs remain unaffected, unless specified to be processed      
  NoRecodeNA = recodeX(x,
                      "new_1"   = c("old_4","old_8"),
                      elseLevel = "other")
)         

# keep some levels, collapse others and reset the reference level
ff <- factor(c("apple","pear","banana","kiwi",
               "mango","peach","grape","plum"))
recodeX(ff, 
       stone=c("peach", "plum"), 
       keep=c("apple","banana"),
       elseLevel = "other", ref="stone")


x <- factor(letters[1:6])

z1 <- recodeX(x, AB=c("a","b"), CD=c("c","d"), elseLevel="none of these")
z2 <- recodeX(x, AB=c("a","b"), CD=c("c","d"), elseLevel=NA)
z3 <- recodeX(x, AB=c("a","b"), CD=c("c","d"), elseLevel=NULL)
z4 <- recodeX(x, AB=c("a","b"), GH=c("g","h"), elseLevel=NA, useEmpty=TRUE)
z5 <- recodeX(x, AB=c("a","b"), GH=c("g","h"), elseLevel=NA, useEmpty=FALSE)

data.frame(z1, z2, z3, z4, z5)

lapply(data.frame(z1, z2, z3, z4, z5), levels)

# empty level GH exists in z4...
table(z4, useNA="ifany")
# and is dropped in z5
table(z5, useNA="ifany")

# use integers to define the groups to collapse
set.seed(1972)
(likert <- factor(sample(1:10, size=15, replace=TRUE),
                  levels=1:10, labels=gettextf("(\%s)", 1:10)))
recodeX(likert, det=1:6, pas=7:8, pro=9:10)

# or directly turned to numeric
recodeX(likert, "1"=1:6, "2"=7:8, "5"=9:10, num=TRUE)

}
\seealso{
\code{\link[=factor]{factor()}}, \code{\link[=levels]{levels()}}, \code{\link[=relevel]{relevel()}},
\code{\link[=reorder]{reorder()}}\cr There's another solution for this problem in the
package \strong{car}.

Other data.recode:
\code{\link[=asBinary]{asBinary()}},
\code{\link[=combLevels]{combLevels()}},
\code{\link[=dummy]{dummy()}},
\code{\link[=mReplace]{mReplace()}},
\code{\link[=nf]{nf()}},
\code{\link[=revCode]{revCode()}},
\code{\link[=stringsAsFactors]{stringsAsFactors()}}
}
\concept{categorization}
\concept{data.recode}
\concept{recoding}
