% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleX.R
\name{sampleX}
\alias{sampleX}
\alias{sampleX.data.frame}
\alias{sampleX.default}
\title{Random Samples and Permutations}
\usage{
sampleX(x, size, replace = FALSE, prob = NULL)

\method{sampleX}{data.frame}(x, size = nrow(x), replace = FALSE, prob = NULL)

\method{sampleX}{default}(x, size, replace = FALSE, prob = NULL)
}
\arguments{
\item{x}{either a vector of one or more elements from which to choose,
or a positive integer, or a data frame whose rows are to be sampled.}

\item{size}{a non-negative integer giving the number of items (or rows)
to choose. If missing, it defaults to the number of elements of \code{x}
(resp. \code{nrow(x)} for data frames), yielding a random permutation.}

\item{replace}{logical; whether sampling is performed with replacement.}

\item{prob}{a vector of probability weights for obtaining the elements
(or rows) being sampled.}
}
\value{
sampled elements in the same structure as \code{x}; for data
frames, a data frame containing the sampled rows.
}
\description{
\code{sampleX} takes a sample of the specified size from the elements
of \code{x}, with or without replacement. It does the same as
\code{\link[=sample]{sample()}} and additionally offers an interface for data
frames, where rows are sampled.
}
\examples{

sampleX(1:10, size = 5)

# random permutation, like sample(x)
sampleX(1:10)

# sample rows of a data frame
sampleX(mtcars, size = 5)

}
\seealso{
\code{\link[=sample]{sample()}}

Other combinatorics:
\code{\link[=combN]{combN()}},
\code{\link[=combPairs]{combPairs()}},
\code{\link[=combSet]{combSet()}},
\code{\link[=pairApply]{pairApply()}},
\code{\link[=permn]{permn()}},
\code{\link[=randGroupSplit]{randGroupSplit()}}
}
\concept{combinatorics}
\concept{sampling}
