% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-and-remove-attributes.R
\name{set-and-remove-attributes}
\alias{set-and-remove-attributes}
\alias{setAttr}
\alias{removeAttr}
\title{Set and Remove Object Attributes}
\usage{
setAttr(x, attrNames, attrValues)

removeAttr(x, attrNames = NULL)
}
\arguments{
\item{x}{Object to modify}

\item{attrNames}{Character vector of attribute names}

\item{attrValues}{Values for the attributes (only for setting)}
}
\value{
Modified object
}
\description{
Convenience helpers to add or remove attributes from an object.
}
\examples{
x <- runif(10)

x <- setAttr(
  x,
  attrNames = c("some_attr", "other_attr"),
  attrValues = c("First attribute", "Second attribute")
)

# remove single attribute
removeAttr(x, "other_attr")

# remove all attributes
removeAttr(x)
}
\seealso{
\code{\link{setNamesX}}, \code{\link{unname}}

Other data.manipulation:
\code{\link[=appendRowNames]{appendRowNames()}},
\code{\link[=appendX]{appendX()}},
\code{\link[=as.array.xtabs]{as.array.xtabs()}},
\code{\link[=columnWrap]{columnWrap()}},
\code{\link[=combLevels]{combLevels()}},
\code{\link[=nf]{nf()}},
\code{\link[=parseSASDatalines]{parseSASDatalines()}},
\code{\link[=recodeX]{recodeX()}},
\code{\link[=recycle]{recycle()}},
\code{\link[=revCode]{revCode()}},
\code{\link[=revX]{revX()}},
\code{\link[=setNamesX]{setNamesX()}},
\code{\link[=sortX]{sortX()}},
\code{\link[=splitX]{splitX()}},
\code{\link[=stringsAsFactors]{stringsAsFactors()}},
\code{\link[=toBaseR]{toBaseR()}},
\code{\link{tolong_towide}},
\code{\link[=untable]{untable()}}
}
\concept{data-manipulation}
\concept{data-structures}
\concept{data.manipulation}
