% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripAttr.R
\name{set_strip_attr}
\alias{set_strip_attr}
\alias{setAttr}
\alias{stripAttr}
\alias{SetAttr}
\title{Remove Attributes from an Object}
\usage{
setAttr(x, attr, attr_val)

stripAttr(x, attr_names = NULL)
}
\arguments{
\item{x}{the object whose attributes should be removed or to which an
attribute should be added.}

\item{attr}{name of a new attribute}

\item{attr_val}{value for the new attribute \code{attr}}

\item{attr_names}{a vector with attribute names, which will be removed.
Leaving the default to \code{NULL} will cause all the attributes to be
deleted.}
}
\value{
the object \code{x} without the attributes contained in attr_names
}
\description{
For convenience we sometimes want to strip some or all attributes in a
oneliner.
}
\examples{

x <- runif(10)
x <- setAttr(x, 
             attr=c("some_attr", "other_attr"),
             attr_val=c("First attribute", "Second attribute"))

# strip only single
stripAttr(x, "other_attr")

# strip all attributes
stripAttr(x)
}
\seealso{
\code{\link{setNamesX}}, \code{\link{unname}}

Other data.manipulation: 
\code{\link{appendRowNames}()},
\code{\link{appendX}()},
\code{\link{as.array.xtabs}()},
\code{\link{columnWrap}()},
\code{\link{combLevels}()},
\code{\link{nf}()},
\code{\link{parseSASDatalines}()},
\code{\link{recodeX}()},
\code{\link{recycle}()},
\code{\link{revCode}()},
\code{\link{revX}()},
\code{\link{setNamesX}()},
\code{\link{sortX}()},
\code{\link{splitX}()},
\code{\link{stringsAsFactors}()},
\code{\link{toBaseR}()},
\code{\link{tolong_towide}},
\code{\link{untable}()}
}
\concept{data-manipulation}
\concept{data-structures}
\concept{data.manipulation}
