% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-remove-keep-attr.R
\name{setAttr-removeAttr-keepAttr}
\alias{setAttr-removeAttr-keepAttr}
\alias{setAttr}
\alias{removeAttr}
\alias{keepAttr}
\title{Set and Remove Object Attributes}
\usage{
setAttr(x, attrNames, attrValues)

removeAttr(x, attrNames = NULL)

keepAttr(x, attrNames)
}
\arguments{
\item{x}{object to modify.}

\item{attrNames}{character vector of attribute names.}

\item{attrValues}{values for the attributes (only for setting). For a
single attribute name, \code{attrValues} is taken as the value itself
(which may be a vector). For several names, supply one value per name;
use a list for non-scalar or mixed-type values.}
}
\value{
modified object.
}
\description{
Convenience helpers to add, remove, or selectively retain attributes
of an object.
}
\examples{
x <- runif(10)

x <- setAttr(
  x,
  attrNames = c("some_attr", "other_attr"),
  attrValues = c("First attribute", "Second attribute")
)

# a single attribute can take a vector value
setAttr(1:10, "dim", c(2, 5))

# several non-scalar values via list
setAttr(1:10, c("dim", "myattr"), list(c(2, 5), "abc"))

# remove single attribute
removeAttr(x, "other_attr")

# remove all attributes
removeAttr(x)

# keep only selected attributes, remove all others
r.lm <- lm(Fertility ~ ., swiss)
keepAttr(r.lm$terms, "class")
}
\seealso{
\code{\link[=setNames]{setNames()}}, \code{\link[=unname]{unname()}}

Other label.attrs:
\code{\link[=label]{label()}},
\code{\link[=renameX]{renameX()}},
\code{\link[=setNamesX]{setNamesX()}}
}
\concept{attribute}
\concept{label.attrs}
\concept{programming}
