% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setLength.R
\name{setLength}
\alias{setLength}
\title{Set the length of a vector, padding or truncating as needed}
\usage{
setLength(x, n, fill = NA)
}
\arguments{
\item{x}{a vector}

\item{n}{target length, a single non-negative whole number}

\item{fill}{value used for newly added elements when \code{x} is
extended (default \code{NA})}
}
\value{
\code{x}, of length \code{n}
}
\description{
Extends \code{x} to length \code{n} by appending \code{fill}, or
truncates it to length \code{n} if it is already longer -- like
\code{length(x) <- n}, but with a configurable fill value instead
of \code{NA}.
}
\examples{
setLength(LETTERS[1:3], 5)
setLength(LETTERS[1:3], 2)
setLength(1:4, 6, fill = 0)

}
\seealso{
Other vector.ops:
\code{\link[=closest]{closest()}},
\code{\link[=coalesceX]{coalesceX()}},
\code{\link[=locf]{locf()}},
\code{\link[=midx]{midx()}},
\code{\link[=moveAvg]{moveAvg()}},
\code{\link[=naIf]{naIf()}},
\code{\link[=naReplace]{naReplace()}},
\code{\link[=nz]{nz()}},
\code{\link[=pairApply]{pairApply()}},
\code{\link[=quot]{quot()}},
\code{\link[=trim]{trim()}},
\code{\link[=vRot]{vRot()}},
\code{\link[=vShift]{vShift()}},
\code{\link[=winsorize]{winsorize()}}
}
\concept{ordering}
\concept{vector.ops}
