% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setNamesX.R
\name{setNamesX}
\alias{setNamesX}
\title{Set the Names in an Object}
\usage{
setNamesX(x, ...)
}
\arguments{
\item{x}{an object for which a names attribute will be meaningful.}

\item{...}{the names to be assigned to the object. This should be a
character vector of names named \code{dimnames}, \code{rownames},
\code{colnames} or \code{names}. Setting \code{rownames=NULL} would remove
existing rownames. All kind of names can be changed at the same time.
Default would be \code{names}. Abbreviations are supported.}
}
\value{
an object of the same sort as object with the new names assigned.
}
\description{
This is a convenience function that sets the names of an object and returns
it including the new names. It is most useful at the end of a function
definition where one is creating the object to be returned and would prefer
not to store it under a name just that the names can be assigned. In
addition to the function \code{\link[=setNames]{setNames()}} in base R the user can
decide, whether rownames, colnames or simply the names are to be set.
}
\details{
A name of length one is recycled to the required extent, which is handy for
blanking out names with \code{""}. Names of any other length must match the
extent of the object exactly; a deviating length is reported as an error
rather than being recycled silently, as an unexpected length is almost
always a miscalculation upstream and duplicated names are hard to debug
later on.
}
\examples{

setNamesX(1:5, names=letters[1:5])

# the default, if no argument names are provided, is "names"
setNamesX(1:5, letters[1:5])

# rownames and columnnames can be set at the same time
setNamesX(matrix(c(1:12), nrow=4), 
         rownames=LETTERS[11:14], colnames=c("perc", "lci", "uci"))
         
# a single name is recycled, so this sets all the names to an empty string
setNamesX(diag(6), rownames="", colnames="")

# any other length must fit, a mismatch is an error
try(setNamesX(matrix(c(1:12), nrow=4), colnames=c("perc", "lci")))

# setting dimnames works as well
tab <- setNamesX(
  as.table(rbind(c(84,43), c(10,92))), 
    dimnames= list(
       dipstick=c("positive","negative"),
       culture=c("positive","negative")))


}
\seealso{
\code{\link[=setNames]{setNames()}}

Other label.attrs:
\code{\link[=label]{label()}},
\code{\link[=renameX]{renameX()}},
\code{\link{setAttr-removeAttr-keepAttr}}
}
\concept{attribute}
\concept{label}
\concept{label.attrs}
