% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toBaseR.R
\name{toBaseR}
\alias{toBaseR}
\alias{toBaseR.tbl_df}
\alias{toBaseR.haven_labelled}
\alias{toBaseR.default}
\title{Back to Basics with Tibbles}
\usage{
toBaseR(x, ...)

\method{toBaseR}{tbl_df}(x, ...)

\method{toBaseR}{haven_labelled}(x, ...)

\method{toBaseR}{default}(x, ...)
}
\arguments{
\item{x}{the object to be converted.}

\item{\dots}{arguments passed on.}
}
\value{
converted object.
}
\description{
Sometimes we might wish for the old days be back and want to work with
familiar objects. This function helps to convert \code{tibbles} to
\code{data.frames} as smoothly as possible.
}
\examples{
# a tibble is rolled back to a plain data.frame
if (requireNamespace("tibble", quietly = TRUE)) {
  tbl <- tibble::as_tibble(head(iris))
  class(toBaseR(tbl))
}

# an object without a method is returned unchanged, with a warning
x <- suppressWarnings(toBaseR(1:3))
identical(x, 1:3)

\donttest{
# labelled data from other statistical packages: needs 'haven' and
# an internet connection, hence the try()
if (requireNamespace("haven", quietly = TRUE)) {
  url <- "http://www.stata.com/videos13/data/webclass.dta"
  d.webclass <- try(toBaseR(haven::read_dta(url)))
}
}

}
\seealso{
Other data.coerce:
\code{\link[=as.array.xtabs]{as.array.xtabs()}},
\code{\link{type-aliases}}
}
\concept{data.coerce}
\concept{type-coercion}
