% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toLongWide.R
\name{tolong_towide}
\alias{tolong_towide}
\alias{toLong}
\alias{toWide}
\title{Reshape a Vector From Long to Wide Shape Or Vice Versa}
\usage{
toLong(x, varnames = NULL, incl.rownames = FALSE)

toWide(x, g, by = NULL, varnames = NULL)
}
\arguments{
\item{x}{the vector to be reshaped}

\item{varnames}{the variable names if not the grouping levels should be
used.}

\item{incl.rownames}{logical (default \code{FALSE}), if set to \code{TRUE} a
column containing the rownames will be appended at the end of the
data.frame.}

\item{g}{the grouping vector to be used for the new columns. The resulting
\code{data.frame} will return one column per grouplevel.}

\item{by}{a vector to be used to merge the pieces of \code{x}. If this is
left to \code{NULL} the pieces will be merged by rownames in the order they
are supplied.}
}
\value{
the reshaped object of \code{data.frame} class
}
\description{
Simple reshaping a vector from long to wide or from wide to long shape by
means of a single factor.
}
\details{
\code{toLong} expects x as a matrix or a data.frame and reshapes it to a
(long) factor representation. \code{toWide} expects the vectors x, g, by,
wheras x being the variable, g the splitting factor and by a vector for
rowwise merging.
}
\examples{

d.x <- read.table(header=TRUE, text="
AA BB CC DD EE FF GG
7.9 18.1 13.3 6.2 9.3 8.3 10.6
9.8 14.0 13.6 7.9 2.9 9.1 13.0
6.4 17.4 16.0 10.9 8.6 11.7 17.5
")

toLong(d.x)

# toWide by row numbers (by = NULL)
toWide(PlantGrowth$weight, PlantGrowth$group)

# To wide aligned by key
set.seed(41)
PlantGrowth$nr <- c(sample(12, 10), sample(12, 10), sample(12, 10))
head(PlantGrowth)

toWide(PlantGrowth$weight, PlantGrowth$group, by=PlantGrowth$nr)

}
\seealso{
\code{\link{reshape}}

Other data.manipulation:
\code{\link[=appendRowNames]{appendRowNames()}},
\code{\link[=appendX]{appendX()}},
\code{\link[=as.array.xtabs]{as.array.xtabs()}},
\code{\link[=columnWrap]{columnWrap()}},
\code{\link[=combLevels]{combLevels()}},
\code{\link[=nf]{nf()}},
\code{\link[=parseSASDatalines]{parseSASDatalines()}},
\code{\link[=recodeX]{recodeX()}},
\code{\link[=recycle]{recycle()}},
\code{\link[=revCode]{revCode()}},
\code{\link[=revX]{revX()}},
\code{\link{set-and-remove-attributes}},
\code{\link[=setNamesX]{setNamesX()}},
\code{\link[=sortX]{sortX()}},
\code{\link[=splitX]{splitX()}},
\code{\link[=stringsAsFactors]{stringsAsFactors()}},
\code{\link[=toBaseR]{toBaseR()}},
\code{\link[=untable]{untable()}}
}
\concept{data-manipulation}
\concept{data-structures}
\concept{data.manipulation}
\keyword{manip}
