% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R
\name{trim}
\alias{trim}
\title{Trim a Vector}
\usage{
trim(x, trim = 0.1, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector to be trimmed.}

\item{trim}{the fraction (0 to 0.5) of observations to be trimmed from each
end of x. Values of trim outside that range (and < 1) are taken as the
nearest endpoint.  If \code{trim} is set to a value >1 it's interpreted as
the number of elements to be cut off at each tail of \code{x}.}

\item{na.rm}{a logical value indicating whether \code{NA} values should be
stripped before the computation proceeds.}
}
\value{
the trimmed vector \code{x}. The indices of the trimmed values will
be attached as attribute named \code{"trim"}.
}
\description{
Clean data by means of trimming, i.e., by omitting outlying observations.
}
\details{
A symmetrically trimmed vector \code{x} with a fraction of trim observations
(resp. the given number) deleted from each end will be returned. If
\code{trim} is set to a value \eqn{\ge 0.5} or to an integer value
\eqn{\ge n/2} then the result will be \code{NA}. The same applies if
\code{x} contains \code{NA}s and \code{na.rm} is \code{FALSE}.
}
\note{
This function is basically an excerpt from the base function
\code{\link[=mean]{mean()}}, which allows the vector \code{x} to be trimmed before
calculating the mean.
}
\examples{

## generate data
set.seed(1234)     # for reproducibility
x <- rnorm(10)     # standard normal
x[1] <- x[1] * 10  # introduce outlier

## Trim data
x
trim(x, trim=0.1)

## Trim fixed number, say cut the 3 extreme elements from each end
trim(x, trim=3)

## check function
s <- sample(10:20)
s.tr <- trim(s, trim = 2)
setequal(c(s[attr(s.tr, "trim")], s.tr), s)


}
\seealso{
Other vector.reshape:
\code{\link[=setLength]{setLength()}},
\code{\link[=vRot]{vRot()}},
\code{\link[=vShift]{vShift()}}
}
\concept{outlier-detection}
\concept{reshape}
\concept{vector.reshape}
