% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vRot.R
\name{vRot}
\alias{vRot}
\title{Rotate a vector}
\usage{
vRot(x, k = 1L)
}
\arguments{
\item{x}{a vector}

\item{k}{integer. Number of positions to rotate (default = 1).}
}
\value{
A vector of the same length as \code{x}, rotated cyclically.
}
\description{
Rotates a vector cyclically to the right by \code{k} positions.
Negative values of \code{k} rotate to the left.
}
\details{
The rotation is cyclic, meaning elements shifted off one end reappear on the other.
}
\examples{
vRot(1:5, 2)
# 4 5 1 2 3

vRot(1:5, -1)
# 2 3 4 5 1

}
\seealso{
Other vector.reshape:
\code{\link[=setLength]{setLength()}},
\code{\link[=trim]{trim()}},
\code{\link[=vShift]{vShift()}}
}
\concept{reshape}
\concept{shift}
\concept{vector.reshape}
