% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vShift.R
\name{vShift}
\alias{vShift}
\title{Shift a vector with NA padding}
\usage{
vShift(x, k = 1L)
}
\arguments{
\item{x}{a vector.}

\item{k}{integer. Number of positions to shift.
Positive values shift to the right, negative values to the left.}
}
\value{
a vector of the same length as \code{x}, shifted with \code{NA} padding.
}
\description{
Shifts a vector to the left or right by \code{k} positions.
Vacated positions are filled with \code{NA}.
}
\details{
Unlike \code{\link[=vRot]{vRot()}}, this function does not wrap elements around.
Elements shifted beyond the vector bounds are discarded.
}
\examples{
vShift(1:5, 2)
# NA NA 1 2 3

vShift(1:5, -2)
# 3 4 5 NA NA

vShift(1:5, 10)
# NA NA NA NA NA

}
\seealso{
Other vector.reshape:
\code{\link[=setLength]{setLength()}},
\code{\link[=trim]{trim()}},
\code{\link[=vRot]{vRot()}}
}
\concept{reshape}
\concept{shift}
\concept{vector.reshape}
