% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allIdentical.R
\name{allIdentical}
\alias{allIdentical}
\title{Test Multiple Objects for Exact Equality}
\usage{
allIdentical(...)
}
\arguments{
\item{\dots}{any \code{R} objects}
}
\description{
The function \code{\link{identical}()} is the safe and reliable way to test
two objects for being exactly equal. But it is restricted to the comparison
of two objects. \code{allIdentical()} allows the input of multiple objects
and returns \code{TRUE} in the case that all of them are exactly equal,
\code{FALSE} in every other case.
}
\details{
The function checks the first object against all others, so if the first
object is identical to the second and to the third, then also the second and
the third are identical. (If A=B and A=C then is B=C)
}
\examples{

A <- LETTERS[1:5]
B <- LETTERS[1:5]
C <- LETTERS[1:5]
D <- LETTERS[1:5]
E <- factor(LETTERS[1:5])

# all ok
allIdentical(A, B, C, D)

# at least one odd man
allIdentical(A, B, C, D, E)

}
\seealso{
\code{\link{identical}()}
}
\author{
Andri Signorell \href{mailto:andri@signorell.net}{andri@signorell.net}
}
\keyword{IO}
\keyword{iteration}
\keyword{logic}
\keyword{programming}
