% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allIdentical.R
\name{allIdentical}
\alias{allIdentical}
\title{Test Whether Multiple Objects Are Identical}
\usage{
allIdentical(...)
}
\arguments{
\item{...}{Objects to compare.}
}
\value{
Logical scalar.
}
\description{
Extends \code{\link{identical}} to more than two objects. Returns
\code{TRUE} if all supplied objects are exactly identical, and
\code{FALSE} otherwise.
}
\examples{
A <- LETTERS[1:5]
B <- LETTERS[1:5]
C <- LETTERS[1:5]
E <- factor(LETTERS[1:5])

allIdentical(A, B, C)        # TRUE
allIdentical(A, B, C, E)     # FALSE

}
\seealso{
\code{\link{identical}}

Other data.inspection:
\code{\link[=allDuplicated]{allDuplicated()}},
\code{\link[=completeColumns]{completeColumns()}},
\code{\link[=countCompCases]{countCompCases()}},
\code{\link[=flags]{flags()}},
\code{\link[=isDichotomous]{isDichotomous()}},
\code{\link[=isNumeric]{isNumeric()}},
\code{\link[=isWholeLike]{isWholeLike()}},
\code{\link[=isZero]{isZero()}}
}
\concept{data-inspection}
\concept{data.inspection}
