% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendRowNames.R
\name{appendRowNames}
\alias{appendRowNames}
\title{Append Rownames to a Data Frame or Matrix}
\usage{
appendRowNames(x, col_name = "rowname", after = 0L, remove_rownames = TRUE)
}
\arguments{
\item{x}{A data.frame or matrix.}

\item{col_name}{Name of the new column containing the row names.}

\item{after}{Position after which the column is inserted.
Default is 0 (first column).}

\item{remove_rownames}{Logical; if TRUE, existing row names are removed.}
}
\value{
A data.frame (or matrix coerced to data.frame) with row names
added as a column.
}
\description{
Adds the row names of a data.frame or matrix as a column.
}
\examples{
dd <- data.frame(x=1:5, y=6:10, z=LETTERS[1:5],
                 row.names = letters[1:5])
appendRowNames(dd)

}
\seealso{
\code{\link{appendX}}

Other data.manipulation: 
\code{\link{appendX}()},
\code{\link{as.array.xtabs}()},
\code{\link{columnWrap}()},
\code{\link{combLevels}()},
\code{\link{nf}()},
\code{\link{parseSASDatalines}()},
\code{\link{recodeX}()},
\code{\link{recycle}()},
\code{\link{revCode}()},
\code{\link{revX}()},
\code{\link{setNamesX}()},
\code{\link{set_strip_attr}},
\code{\link{sortX}()},
\code{\link{splitX}()},
\code{\link{stringsAsFactors}()},
\code{\link{toBaseR}()},
\code{\link{tolong_towide}},
\code{\link{untable}()}
}
\concept{data-manipulation}
\concept{data.manipulation}
