% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendRowNames.R
\name{appendRowNames}
\alias{appendRowNames}
\title{Append Rownames to a Data Frame}
\usage{
appendRowNames(x, col_name = "rowname", after = 0, remove_rownames = TRUE)
}
\arguments{
\item{x}{a data.frame}

\item{col_name}{the name of the new inserted column containing the rownames.}

\item{after}{a subscript, after which the values are to be appended.  If
missing the rownames will be inserted as first column.}

\item{remove_rownames}{logical defining if the existing rownames should be
removed.  Default is \code{TRUE}.}
}
\value{
the object x with appended rownames
}
\description{
Append rownames to a data.frame as first column.
}
\examples{

(dd <- data.frame(x=1:5, y=6:10, z=LETTERS[1:5], 
                  row.names = letters[1:5]))
appendRowNames(dd)

}
\seealso{
\code{\link{appendX}}
}
