% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendRowNames.R
\name{appendRowNames}
\alias{appendRowNames}
\title{Append Rownames to a Data Frame or Matrix}
\usage{
appendRowNames(x, colName = "rowname", after = 0L, removeRownames = TRUE)
}
\arguments{
\item{x}{A data.frame or matrix.}

\item{colName}{Name of the new column containing the row names.}

\item{after}{Position after which the column is inserted.
Default is 0 (first column).}

\item{removeRownames}{Logical; if TRUE, existing row names are removed.}
}
\value{
A data.frame (or matrix coerced to data.frame) with row names
added as a column.
}
\description{
Adds the row names of a data.frame or matrix as a column.
}
\examples{
dd <- data.frame(x=1:5, y=6:10, z=LETTERS[1:5],
                 row.names = letters[1:5])
appendRowNames(dd)

}
\seealso{
\code{\link{append}}

Other data.manipulation:
\code{\link[=appendX]{appendX()}},
\code{\link[=as.array.xtabs]{as.array.xtabs()}},
\code{\link[=columnWrap]{columnWrap()}},
\code{\link[=combLevels]{combLevels()}},
\code{\link[=nf]{nf()}},
\code{\link[=parseSASDatalines]{parseSASDatalines()}},
\code{\link[=recodeX]{recodeX()}},
\code{\link[=recycle]{recycle()}},
\code{\link[=revCode]{revCode()}},
\code{\link[=revX]{revX()}},
\code{\link{set-and-remove-attributes}},
\code{\link[=setNamesX]{setNamesX()}},
\code{\link[=sortX]{sortX()}},
\code{\link[=splitX]{splitX()}},
\code{\link[=stringsAsFactors]{stringsAsFactors()}},
\code{\link[=toBaseR]{toBaseR()}},
\code{\link{tolong_towide}},
\code{\link[=untable]{untable()}}
}
\concept{data-manipulation}
\concept{data.manipulation}
