% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendX.R
\name{appendX}
\alias{appendX}
\alias{appendX.matrix}
\alias{appendX.data.frame}
\alias{appendX.default}
\alias{appendX.TOne}
\title{Append Elements to Objects}
\usage{
appendX(x, values, after = NULL, ...)

\method{appendX}{default}(x, values, after = NULL, ...)

\method{appendX}{matrix}(x, values, after = NULL, rows = FALSE, names = NULL, ...)

\method{appendX}{data.frame}(x, values, after = NULL, rows = FALSE, names = NULL, ...)

\method{appendX}{TOne}(x, values, after = NULL, rows = TRUE, names = NULL, ...)
}
\arguments{
\item{x}{object for the elements to be inserted}

\item{values}{the elements to be inserted}

\item{after}{a subscript, after which the values are to be appended. If it's
missing the values will be appended after the last element (or column/row).}

\item{\dots}{further arguments (not used here)}

\item{rows}{logical, defining if vector should be added as row or as column.
Default is column (\code{rows}=\code{FALSE}).}

\item{names}{the dimension names for the inserted elements(s)}
}
\value{
An object containing the values in x with the elements of values
appended after the specified element of x.
}
\description{
Append elements to a number of various objects as vectors, matrices,
data.frames and lists. In a matrix either rows or columns can be inserted at
any position. In data frames any vectors can be inserted. \code{values} will
be recycled to the necessary length.
}
\details{
The vector \code{x} will be recycled to a length of the next multiple of the
number of rows (or columns) of the matrix \code{m} and will be inserted such
that the first inserted row (column) has the index \code{i}. If the dimnames
are given, they will be used no matter if the matrix m has already dimnames
defined or not.
}
\examples{


appendX(1:5, 0:1, after = 3)    # the same as append

# Insert columns and rows
x <- matrix(runif(25), 5)

appendX(x, values=1:10, after=2, names = c("X","Y"))
appendX(x, values=1:10, after=2)

appendX(x, values=1:10, after=2, names = c("X","Y"))
appendX(x, values=1:10, after=2)

# append to a data.frame
d.frm <- data.frame("id"   = c(1,2,3),
                    "code" = c("AAA", "BBB", "CCC"),
                    "val"  = c(111, 222, 333))
z <- c(10, 20, 30)

appendX(d.frm, z, after=2, names="ZZZ")

}
\seealso{
\code{\link{rbind}}, \code{\link{cbind}}, \code{\link{append}}
}
\author{
Andri Signorell \href{mailto:andri@signorell.net}{andri@signorell.net}
}
\keyword{manip}
