% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.matrix.xtabs.R
\name{as.array.xtabs}
\alias{as.array.xtabs}
\alias{as.matrix.xtabs}
\title{Coerce \code{xtabs} Object to Array or Matrix}
\usage{
\method{as.array}{xtabs}(x, ...)

\method{as.matrix}{xtabs}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"xtabs"}.}

\item{...}{ignored.}
}
\value{
an array (or matrix in the two-dimensional case) with no
additional classes.
}
\description{
Converts an object of class \code{"xtabs"} to a plain array or matrix by
dropping all additional classes such as \code{"xtabs"} and \code{"table"},
along with the \code{call} attribute.
}
\examples{
xt <- xtabs(~ cyl + gear, data = mtcars)
class(as.matrix(xt))
# "matrix" "array"

}
\seealso{
Other data.coerce:
\code{\link[=toBaseR]{toBaseR()}},
\code{\link{type-aliases}}
}
\concept{data.coerce}
\concept{table}
\concept{type-coercion}
