% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.matrix.xtabs.R
\name{as.array.xtabs}
\alias{as.array.xtabs}
\title{Coerce \code{xtabs} Object to Array or Matrix}
\usage{
\method{as.array}{xtabs}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"xtabs"}.}

\item{...}{Ignored.}
}
\value{
A matrix with no additional classes.
}
\description{
Converts an object of class \code{"xtabs"} to a plain matrix by
dropping all additional classes such as \code{"xtabs"} and \code{"table"}.
}
\examples{
xt <- xtabs(~ cyl + gear, data = mtcars)
class(as.matrix(xt))
# "matrix"

}
\seealso{
Other data.manipulation:
\code{\link[=appendRowNames]{appendRowNames()}},
\code{\link[=appendX]{appendX()}},
\code{\link[=columnWrap]{columnWrap()}},
\code{\link[=combLevels]{combLevels()}},
\code{\link[=nf]{nf()}},
\code{\link[=parseSASDatalines]{parseSASDatalines()}},
\code{\link[=recodeX]{recodeX()}},
\code{\link[=recycle]{recycle()}},
\code{\link[=revCode]{revCode()}},
\code{\link[=revX]{revX()}},
\code{\link{set-and-remove-attributes}},
\code{\link[=setNamesX]{setNamesX()}},
\code{\link[=sortX]{sortX()}},
\code{\link[=splitX]{splitX()}},
\code{\link[=stringsAsFactors]{stringsAsFactors()}},
\code{\link[=toBaseR]{toBaseR()}},
\code{\link{tolong_towide}},
\code{\link[=untable]{untable()}}
}
\concept{data-manipulation}
\concept{data-structures}
\concept{data.manipulation}
\concept{table-manipulation}
