% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.array.xtabs.R
\name{as.array.xtabs}
\alias{as.array.xtabs}
\alias{as.matrix.xtabs}
\title{Coerce \code{xtabs} Object to Array or Matrix}
\usage{
\method{as.array}{xtabs}(x, ...)

\method{as.matrix}{xtabs}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"xtabs"}.}

\item{...}{Further arguments (currently ignored).}
}
\value{
An array (or matrix) with the same data, dimensions, and dimnames as
\code{x}, but without the \code{"xtabs"} class.
}
\description{
Converts an object of class \code{"xtabs"} to a plain array or matrix by
removing class-specific attributes.
}
\details{
\code{xtabs} objects are internally arrays with additional class and metadata.
This method strips these attributes without altering the underlying data.
}
\examples{
x <- xtabs(Freq ~ ., data = as.data.frame(Titanic))
class(x)

m <- as.matrix(x)
class(m)

}
\seealso{
\code{\link{xtabs}}, \code{\link{as.array}}, \code{\link{as.matrix}}
}
