% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest.R
\name{closest}
\alias{closest}
\title{Find the Closest Value}
\usage{
closest(x, a, which = FALSE, na.rm = FALSE)
}
\arguments{
\item{x}{the vector to be searched in}

\item{a}{the reference value}

\item{which}{a logical value defining if the index position or the value
should be returned. By default will the value be returned.}

\item{na.rm}{a logical value indicating whether \code{NA} values should be
stripped before the computation proceeds.}
}
\value{
the value or index in x which is closest to a
}
\description{
Find the closest value(s) of a number in a vector x. Multiple values will be
reported, if the differences are the same or if there are duplicates of the
same value.
}
\examples{


# basic
set.seed(8)
x <- runif(10) * 10
closest(x, 3.1)
sort(x)

y <- sample(10, size=10, replace=TRUE)
# multiple observations of the same closest value 
closest(y, a=6)
# get the relevant positions
closest(y, a=6, which=TRUE)

# two different values having the same distance
closest(c(2, 3, 4, 5), a=3.5)

# vectorize "a"
closest(c(2, 3, 4, 5), a=c(3.1, 3.9))

# vectorize "which"
closest(c(2, 3, 4, 5), a=3.1, which=c(FALSE, TRUE))

# vectorize both
closest(c(2, 3, 4, 5), a=c(3.1, 3.9), which=c(FALSE, TRUE))

}
\seealso{
\code{\link{which}}
}
