% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/columnWrap.R
\name{columnWrap}
\alias{columnWrap}
\title{Column Wrap}
\usage{
columnWrap(x, width = NULL)
}
\arguments{
\item{x}{the matrix with one row}

\item{width}{integer, the width of the columns in characters}
}
\value{
a character matrix
}
\description{
Wraps text in a character matrix so, that it's displayed over more than one
line.
}
\details{
A data.frame containing character columns with long texts is often wrapped
by columns. This can lead to a loss of overview. \code{columnWrap()} wraps the
lines within the columns.
}
\examples{

print(columnWrap("This is a very long text for a table", 12))
}
\seealso{
\code{\link[=strwrap]{strwrap()}}

Other data.manipulation:
\code{\link[=appendRowNames]{appendRowNames()}},
\code{\link[=appendX]{appendX()}},
\code{\link[=as.array.xtabs]{as.array.xtabs()}},
\code{\link[=combLevels]{combLevels()}},
\code{\link[=nf]{nf()}},
\code{\link[=parseSASDatalines]{parseSASDatalines()}},
\code{\link[=recodeX]{recodeX()}},
\code{\link[=recycle]{recycle()}},
\code{\link[=revCode]{revCode()}},
\code{\link[=revX]{revX()}},
\code{\link{set-and-remove-attributes}},
\code{\link[=setNamesX]{setNamesX()}},
\code{\link[=sortX]{sortX()}},
\code{\link[=splitX]{splitX()}},
\code{\link[=stringsAsFactors]{stringsAsFactors()}},
\code{\link[=toBaseR]{toBaseR()}},
\code{\link{tolong_towide}},
\code{\link[=untable]{untable()}}
}
\concept{data-manipulation}
\concept{data-structures}
\concept{data.manipulation}
\keyword{print}
