% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combN.R
\name{combN}
\alias{combN}
\title{Number of Combinations of a Set}
\usage{
combN(n, m, replace = FALSE, ordered = FALSE)
}
\arguments{
\item{n}{number of elements from which to choose.}

\item{m}{number of elements to choose. For \code{combSet} can \code{m} be a
numeric vector too.}

\item{replace}{logical. Should repetition of the same element be allowed?
Defaults to \code{FALSE}}

\item{ordered}{logical. Does the order matter? Default is \code{FALSE}.}
}
\value{
a numeric value
}
\description{
Return the number of combinations with and without replacement and order.
}
\examples{
n <- 5; m <- 2
combN(n, m, replace=TRUE, ordered=FALSE)
combN(n, m, replace=TRUE, ordered=TRUE)
combN(n, m, replace=FALSE, ordered=TRUE)
combN(n, m, replace=FALSE, ordered=FALSE)

}
\seealso{
\code{\link{combPairs}},
\code{\link{combn}}, \code{\link{choose}}, \code{\link{factorial}},
\cr \code{vignette("Combinatorics")}

Other combinatorics:
\code{\link[=combPairs]{combPairs()}},
\code{\link[=combSet]{combSet()}},
\code{\link[=permn]{permn()}},
\code{\link[=randGroupSplit]{randGroupSplit()}},
\code{\link[=sampleX]{sampleX()}},
\code{\link[=unwhich]{unwhich()}}
}
\concept{combinatorics}
