% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeColumns.R
\name{completeColumns}
\alias{completeColumns}
\title{Identify Columns Without Missing Values}
\usage{
completeColumns(x, which = TRUE)
}
\arguments{
\item{x}{A data.frame or list-like object whose elements are checked
for missing values.}

\item{which}{Logical. If \code{TRUE} (default), the function returns
the names of elements without missing values. If \code{FALSE}, a
logical vector is returned, where each element corresponds to an
element in \code{x} and indicates whether it is complete.}
}
\value{
If \code{which = TRUE}, a character vector with the names of all
complete elements.

If \code{which = FALSE}, a logical vector of length \code{length(x)},
where \code{TRUE} indicates that the corresponding element contains
no missing values.
}
\description{
This function checks each element of a data frame or list-like object
for missing values (\code{NA}) and identifies those that are completely
observed, i.e., contain no missing entries.
}
\details{
Depending on the argument \code{which}, the function either returns
the names of such elements or a logical vector indicating completeness
for each element.

An element is considered \emph{complete} if it contains zero missing values.
Internally, the function uses \code{\link{anyNA}} to detect missing values.
}
\examples{
# Names of columns without missing values
completeColumns(d.pizza)

# Logical vector indicating completeness
completeColumns(d.pizza, which = FALSE)

}
\seealso{
\code{\link{anyNA}}, \code{\link{is.na}},
\code{\link{na.omit}}, \code{\link{complete.cases}}

Other data.inspection: 
\code{\link{allDuplicated}()},
\code{\link{allIdentical}()},
\code{\link{countCompCases}()},
\code{\link{flags}()},
\code{\link{isDichotomous}()},
\code{\link{isNumeric}()},
\code{\link{isWholeLike}()},
\code{\link{isZero}()}
}
\concept{data-inspection}
\concept{data.inspection}
\concept{missing-data}
