% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countCompCases.R
\name{countCompCases}
\alias{countCompCases}
\title{Count Complete Cases}
\usage{
countCompCases(x)
}
\arguments{
\item{x}{a data.frame containing the data.}
}
\value{
A list with three elements. The first gives the number of rows, the
second the number of complete cases for the whole data frame. The third
element \code{tab} contains the data for the single variables.
}
\description{
Return for each variable of a data frame the number of missing values and
the complete cases to be expected if this variable would be omitted.
}
\examples{
countCompCases(airquality)

}
\seealso{
\code{\link[aurora]{plotMiss}},
\code{\link{complete.cases}}, \code{\link{is.na}}, \code{\link{na.omit}}

Other data.inspection:
\code{\link[=allDuplicated]{allDuplicated()}},
\code{\link[=allIdentical]{allIdentical()}},
\code{\link{between}},
\code{\link[=compareDataFrames]{compareDataFrames()}},
\code{\link[=completeColumns]{completeColumns()}},
\code{\link[=flags]{flags()}},
\code{\link[=isDichotomous]{isDichotomous()}},
\code{\link[=isEuclid]{isEuclid()}},
\code{\link[=isNumeric]{isNumeric()}},
\code{\link[=isURL]{isURL()}},
\code{\link[=isWholeLike]{isWholeLike()}},
\code{\link[=isZero]{isZero()}}
}
\concept{data.inspection}
\concept{missing-value}
