% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countCompCases.R
\name{countCompCases}
\alias{countCompCases}
\alias{print.countCompCases}
\title{Count Complete Cases}
\usage{
countCompCases(x)
}
\arguments{
\item{x}{a data.frame containg the data.}
}
\value{
A list with three elements. The first gives the number of rows, the
second the number of complete cases for the whole data frame. The third
element \code{tab} contains the data for the single variables.
}
\description{
Return for each variable of a data frame the number of missing values and
the complete cases to be expected if this variable would be omitted.
}
\details{
Count Complete Cases

Return for each variable of a data frame the number of missing values and
the complete cases to be expected if this variable would be omitted.
}
\examples{
countCompCases(airquality)


}
\seealso{
\code{\link[aurora]{plotMiss}}, \code{\link{completeColumns}},
\code{\link{complete.cases}}, \code{\link{is.na}}, \code{\link{na.omit}}
}
