% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digitSum.R
\name{digitSum}
\alias{digitSum}
\title{Digit sum for integer values}
\usage{
digitSum(x)
}
\arguments{
\item{x}{An integer vector.}
}
\value{
An integer vector containing the digit sums.
}
\description{
Computes the sum of digits for integer inputs. Negative values are handled
by taking the absolute value.
}
\details{
The function only accepts integer values. If non-integer numerics are supplied,
an error is thrown. Missing values (\code{NA}) are propagated.
}
\examples{
digitSum(124L)
digitSum(c(10L, 99L, -1234L))

}
