% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisors.R
\name{divisors}
\alias{divisors}
\title{Calculate Divisors}
\usage{
divisors(x)
}
\arguments{
\item{x}{vector of positive whole numbers for which the divisors are to
be returned.}
}
\value{
an integer vector containing the proper divisors if \code{x} is a
single number, otherwise a named list of such vectors.
}
\description{
Calculate the proper divisors of positive natural numbers.
}
\details{
Divisibility is a mathematical relationship between two integers. An integer
is divisible by another integer if there is no remainder in the division.
This function returns the \emph{proper} divisors of \code{x}, i.e. all
positive divisors excluding \code{x} itself. The number 11 is prime and has
only the proper divisor 1, whereas the number 12 has the proper divisors
1, 2, 3, 4 and 6. In elementary number theory, the concept of divisibility
is limited to natural numbers. The number of proper divisors can be
determined with the function \code{\link{length}()}.
}
\examples{

divisors(786)

divisors(c(145, 786))

}
\seealso{
Other number.theory:
\code{\link{GCD-LCM}},
\code{\link[=digitSum]{digitSum()}},
\code{\link[=factorize]{factorize()}},
\code{\link[=fibonacci]{fibonacci()}},
\code{\link[=isOdd]{isOdd()}},
\code{\link[=isPrime]{isPrime()}},
\code{\link[=primes]{primes()}}
}
\concept{number-theory}
\concept{number.theory}
\concept{numerical-methods}
