% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisors.R
\name{divisors}
\alias{divisors}
\title{Calculate Divisors}
\usage{
divisors(x)
}
\arguments{
\item{x}{integer number for which the divisors are to be returned}
}
\value{
an integer vector containg the divisors
}
\description{
Calculate divisors of positive natural numbers.
}
\details{
Divisibility is a mathematical relationship between two integers. An integer
is divisible by another integer if there is no remainder in the division.
The number 11 has only two divisors: 1 and the number 11 itself, whereas the
number 12 has many divisors: 1, 2, 3, 4, 6 and 12.  In elementary number
theory, the concept of divisibility is limited to natural numbers.  The
number of its divisors can be determined with the function
\code{\link{length}()}.
}
\examples{

divisors(c(145, 786))
}
\seealso{
Other math.utils:
\code{\link[=crossProd]{crossProd()}},
\code{\link[=crossProdN]{crossProdN()}},
\code{\link[=dotProd]{dotProd()}},
\code{\link{gcd_lcm}},
\code{\link[=percentRank]{percentRank()}},
\code{\link{precision}},
\code{\link[=ptInPoly]{ptInPoly()}},
\code{\link[=roundTo]{roundTo()}},
\code{\link[=unirootAll]{unirootAll()}}
}
\author{
Andri Signorell \href{mailto:andri@signorell.net}{andri@signorell.net}
}
\concept{math.utils}
\concept{mathematics}
\concept{number-theory}
