% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{dummy}
\alias{dummy}
\title{Generate Dummy Codes for a Factor}
\usage{
dummy(
  x,
  method = c("treatment", "sum", "helmert", "poly", "full"),
  base = 1,
  levels = NULL
)
}
\arguments{
\item{x}{Factor or vector of classes.}

\item{method}{Character string specifying the contrast method. One of
\code{"treatment"}, \code{"sum"}, \code{"helmert"}, \code{"poly"}, or \code{"full"}.
Abbreviations are accepted.}

\item{base}{Integer or character string specifying the baseline group for
treatment contrasts.}

\item{levels}{Optional character vector specifying the possible levels of
\code{x}. If \code{NULL}, levels are inferred by \code{factor(x)}.}
}
\value{
A matrix with dummy codes. The number of rows equals \code{length(x)}.
For \code{method = "full"}, the number of columns equals the number of levels.
Otherwise, the number of columns equals the number of levels minus one.

The returned matrix has an attribute \code{"base"} containing the baseline level
for non-full coding, and \code{NA} for full coding.
}
\description{
Generate a matrix of dummy codes, also known as class indicators, for a
factor or class vector.
}
\details{
The argument \code{method} controls the contrast coding. The option \code{"full"}
returns one indicator column for each level of \code{x}. This full-rank coding is
usually redundant for \code{lm()} and related modelling functions.
}
\examples{
x <- c("red", "blue", "green", "blue", "green", "red", "red", "blue")
dummy(x)
dummy(x, base = 2)
dummy(x, method = "sum")

y <- c("Max", "Max", "Max", "Max", "Max", "Bill", "Bill", "Bill")
dummy(y)
dummy(y, base = "Max")
dummy(y, base = "Max", method = "full")

# Revert full dummy coding
m <- dummy(y, method = "full")
apply(m, 1, function(z) colnames(m)[z == 1])

# Revert treatment dummy coding
m <- dummy(y)
apply(
  m,
  1,
  function(z) ifelse(sum(z) == 0, attr(m, "base"), colnames(m)[z == 1])
)

}
\seealso{
\code{model.frame}, \code{contrasts}, \code{stats::contr.treatment},
\code{stats::contr.sum}, \code{stats::contr.helmert}, \code{stats::contr.poly}

Other data.manipulation:
\code{\link[=appendEnum]{appendEnum()}},
\code{\link[=appendRowNames]{appendRowNames()}},
\code{\link[=appendX]{appendX()}},
\code{\link[=as.array.xtabs]{as.array.xtabs()}},
\code{\link[=asBinary]{asBinary()}},
\code{\link[=columnWrap]{columnWrap()}},
\code{\link[=combLevels]{combLevels()}},
\code{\link[=nf]{nf()}},
\code{\link[=recodeX]{recodeX()}},
\code{\link[=recycle]{recycle()}},
\code{\link[=renameX]{renameX()}},
\code{\link[=revCode]{revCode()}},
\code{\link[=revX]{revX()}},
\code{\link{setAttr-removeAttr-keepAttr}},
\code{\link[=setNamesX]{setNamesX()}},
\code{\link[=sortX]{sortX()}},
\code{\link[=splitAt]{splitAt()}},
\code{\link[=splitX]{splitX()}},
\code{\link[=stringsAsFactors]{stringsAsFactors()}},
\code{\link[=toBaseR]{toBaseR()}}
}
\concept{categorization}
\concept{data.manipulation}
\concept{dummy-coding}
