% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize.R
\name{factorize}
\alias{factorize}
\title{Prime Factorization of Integers}
\usage{
factorize(n)
}
\arguments{
\item{n}{vector of positive whole numbers to factorize.}
}
\value{
A named \code{\link{list}} of the same length as \code{n}, each
element a 2-column matrix with column \code{"p"} the prime factors and
column \code{"m"} their respective exponents (or multiplicities), i.e., for a
prime number \code{n}, the resulting matrix is \code{cbind(p = n, m = 1)}.
}
\description{
Compute the prime factorization(s) of integer(s) \code{n}.
Prime factorization of integer(s) works via \code{\link{primes}},
currently in a cheap way, sub-optimal for large composite \code{n}.
}
\details{
For factorization of moderately or really large numbers, see the \pkg{gmp}
package, and its \code{gmp::factorize()} (which is ~20x faster!).
}
\examples{

factorize(47)
factorize(seq(101, 120, by=2))

}
\seealso{
Other number.theory:
\code{\link[=digitSum]{digitSum()}},
\code{\link[=divisors]{divisors()}},
\code{\link[=fibonacci]{fibonacci()}},
\code{\link{gcd_lcm}},
\code{\link[=isOdd]{isOdd()}},
\code{\link[=isPrime]{isPrime()}},
\code{\link{numeric-conversions}},
\code{\link[=primes]{primes()}}
}
\concept{number-theory}
\concept{number.theory}
