% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDownload.R
\name{findDownload}
\alias{findDownload}
\title{Locate a File in the Downloads Directory}
\usage{
findDownload(file)
}
\arguments{
\item{file}{Character string. Name of the file.}
}
\value{
A character string giving the full path to the file.
}
\description{
Returns the full path to a file located in the user's Downloads directory.
}
\details{
The function resolves the path to the user's Downloads directory using
an internal helper and appends \code{file}. It does not perform any
downloading; it only locates files that already exist locally.

If the file does not exist, an error is thrown.
}
\examples{
\dontrun{
# Locate a file in Downloads
findDownload("data.csv")
}

}
\seealso{
Other file.utils:
\code{\link[=buildPath]{buildPath()}},
\code{\link[=fileExistURL]{fileExistURL()}},
\code{\link[=parseSASDatalines]{parseSASDatalines()}},
\code{\link[=pdfManual]{pdfManual()}},
\code{\link[=splitPath]{splitPath()}}
}
\concept{file.utils}
\concept{programming}
