% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flags.R
\name{flags}
\alias{flags}
\title{Extract Dichotomous (Binary) Variables}
\usage{
flags(x, strict = FALSE, na.rm = FALSE, output = "data")
}
\arguments{
\item{x}{A data frame or matrix.}

\item{strict}{Logical.

If \code{TRUE}, only variables with exactly two
distinct values are considered dichotomous.

If \code{FALSE}, variables with one or two
distinct values are allowed.

Default is \code{FALSE}.}

\item{na.rm}{Logical.

Should missing values be ignored when checking
for dichotomous variables?

Default is \code{FALSE}.}

\item{output}{Character string specifying the output representation.

One of:

\describe{
\item{\code{\"data\"}}{
Subset of \code{x} containing only dichotomous variables.
}

\if{html}{\out{<div class="sourceCode">}}\preformatted{\\item\{\code{\"names\"}\}\{
  Names of dichotomous variables.
\}

\\item\{\code{\"index\"}\}\{
  Column indices of dichotomous variables.
\}

\\item\{\code{\"logical\"}\}\{
  Logical vector indicating dichotomous variables.
\}
}\if{html}{\out{</div>}}

}

Default is \code{\"data\"}.}
}
\value{
Depending on \code{output}:

\itemize{
\item \code{\"data\"}: data frame or matrix
\item \code{\"names\"}: character vector
\item \code{\"index\"}: integer vector
\item \code{\"logical\"}: logical vector
}
}
\description{
Identify and extract dichotomous (binary) variables from a data frame
or matrix using \code{isDichotomous()}.
}
\details{
Variables with only missing values are not considered dichotomous
when \code{na.rm = FALSE}.

When \code{na.rm = TRUE}, such variables are treated as empty vectors
and are considered dichotomous only if \code{strict = FALSE}.

Internally, variables with indeterminate dichotomous status
(i.e. \code{NA} returned by \code{isDichotomous()})
are treated as non-dichotomous for filtering purposes.
}
\examples{
dat <- data.frame(
  a = c(0, 1, 1, 0),
  b = c(1, 2, 3, 4),
  c = c(TRUE, FALSE, TRUE, TRUE),
  d = c(NA, NA, NA, NA)
)

flags(dat)

# effect of na.rm
flags(dat, na.rm = TRUE)

# return variable names
flags(dat, output = "names")

# return column indices
flags(dat, output = "index")

}
\seealso{
\code{\link{isDichotomous}}

Other data.inspection:
\code{\link[=allDuplicated]{allDuplicated()}},
\code{\link[=allIdentical]{allIdentical()}},
\code{\link[=completeColumns]{completeColumns()}},
\code{\link[=countCompCases]{countCompCases()}},
\code{\link[=isDichotomous]{isDichotomous()}},
\code{\link[=isNumeric]{isNumeric()}},
\code{\link[=isWholeLike]{isWholeLike()}},
\code{\link[=isZero]{isZero()}}
}
\concept{data-inspection}
\concept{data-manipulation}
\concept{data.inspection}
\concept{factor-handling}
