% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funArgs.R
\name{funArgs}
\alias{funArgs}
\title{List all arguments in a function.}
\usage{
funArgs(fun, package = NULL, sort = FALSE)
}
\arguments{
\item{fun}{the name of the function}

\item{package}{the name of the package which contains the function}

\item{sort}{logical (default \code{FALSE}) should the arguments be alphabetically sorted?}
}
\description{
List all arguments in a function.
}
\examples{

funArgs("combN")
funArgs("combN")$value

cat(attr(funArgs("combN"), "string"))

}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) \emph{The
New S Language}. Wadsworth & Brooks/Cole.
}
\seealso{
\code{\link{ls}}, \code{\link{ls.str}}, \code{\link{lsf.str}}

Other pkg.introspection:
\code{\link[=funCalls]{funCalls()}},
\code{\link[=funKeywords]{funKeywords()}},
\code{\link[=funList]{funList()}},
\code{\link[=getRdLabels]{getRdLabels()}},
\code{\link[=mergeArgs]{mergeArgs()}}
}
\concept{data-inspection}
\concept{package-utilities}
\concept{pkg.introspection}
