% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funArgs.R
\name{funArgs}
\alias{funArgs}
\title{List All Arguments in a Function}
\usage{
funArgs(name, sort = FALSE)
}
\arguments{
\item{name}{the name of the function}

\item{sort}{logical (default \code{FALSE}) should the arguments be alphabetically sorted?}
}
\description{
List all arguments in a function.
}
\examples{

funArgs("combN")
funArgs("combN")$value

cat(attr(funArgs("combN"), "string"))

}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) \emph{The
New S Language}. Wadsworth & Brooks/Cole.
}
\seealso{
\code{\link{ls}}, \code{\link{ls.str}}, \code{\link{lsf.str}}

Other pkg.introspection: 
\code{\link{funCalls}()},
\code{\link{funKeywords}()},
\code{\link{funList}()},
\code{\link{getRdLabels}()},
\code{\link{inDots}()},
\code{\link{mergeArgs}()},
\code{\link{strX}()}
}
\concept{data-inspection}
\concept{package-utilities}
\concept{pkg.introspection}
