% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funCalls.R
\name{funCalls}
\alias{funCalls}
\title{List Calls Used in Function}
\usage{
funCalls(name, package = NULL, sort = FALSE)
}
\arguments{
\item{name}{the name of the function}

\item{package}{the name of the package}

\item{sort}{logical (default \code{FALSE}) should the arguments be alphabetically sorted?}
}
\description{
For screening purposes it can be useful to get a list of all function calls
our function may depend on. \code{funCalls()} parses the function
source and return all found function calls grouped by their package.
}
\note{
Based on code by Nicholas Cooper.
}
\examples{

funCalls("combN", package="bedrock")

}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) \emph{The
New S Language}. Wadsworth & Brooks/Cole.
}
\seealso{
\code{\link{ls}}, \code{\link{ls.str}}, \code{\link{lsf.str}}

Other topic.systemTools: 
\code{\link{buildPath}()},
\code{\link{funArgs}()},
\code{\link{funList}()}
}
\concept{Infrastructure}
\concept{Programming Utilities}
\concept{topic.systemTools}
