% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funKeywords.R
\name{funKeywords}
\alias{funKeywords}
\title{List Keywords For R Manual Pages}
\usage{
funKeywords(topic)
}
\arguments{
\item{topic}{optional, object or man page topic.}
}
\value{
if \code{topic} is missing, the R keywords documentation file is
opened for display via \code{\link[base]{file.show}}, invisibly
returning \code{NULL}. Otherwise, a character vector of topic names
whose keywords match \code{topic}.
}
\description{
List the keywords for specific R man pages or return a list of valid R
keywords.
}
\details{
If \code{topic} is provided, return a list of the Keywords associated with
\code{topic}.  Otherwise, display the list of valid R Keywords from the R
doc/Keywords file.
}
\note{
Substantially based on the \code{keywords()} function from the
\pkg{gtools} package by Gregory R. Warnes, with minor adaptations by the
package author.
}
\examples{
\donttest{
## Show all valid R Keywords
funKeywords()

## Show Keywords associated with the 'merge' function
funKeywords(merge)
funKeywords("merge")
}

}
\seealso{
\code{\link[utils]{help}}

Other pkg.funinfo:
\code{\link[=funArgs]{funArgs()}},
\code{\link[=funCalls]{funCalls()}},
\code{\link[=funList]{funList()}},
\code{\link[=rdLabels]{rdLabels()}},
\code{\link[=rdTitle]{rdTitle()}}
}
\concept{introspection}
\concept{pkg.funinfo}
\concept{programming}
\keyword{documentation}
