% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funKeywords.R
\name{funKeywords}
\alias{funKeywords}
\title{List Keywords For R Manual Pages}
\usage{
funKeywords(topic)
}
\arguments{
\item{topic}{optional, object or man page topic}
}
\description{
List the keywords for specific R man pages or return a list of valid R
keywords.
}
\details{
If \code{topic} is provided, return a list of the Keywords associated with
\code{topic}.  Otherwise, display the list of valid R Keywords from the R
doc/Keywords file.
}
\examples{
\donttest{
## Show all valid R Keywords
funKeywords()

## Show Keywords associated with the 'merge' function
funKeywords(merge)
funKeywords("merge")
}


}
\seealso{
\code{\link[utils]{help}}
}
\author{
Gregory R. Warnes \email{greg@warnes.net}
}
\keyword{documentation}
