% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funList.R
\name{funList}
\alias{funList}
\title{List Objects, Functions Or Data in a Package}
\usage{
funList(package, exported = TRUE)
}
\arguments{
\item{package}{the name of the package}

\item{exported}{logical (default \code{TRUE}) should only exported functions be listed?}
}
\description{
List all the objects, functions or data in a package.
}
\details{
This is just a wrapper for \code{\link{ls}}, \code{\link{ls.str}} and
\code{\link{lsf.str}} with the appropriate arguments (as I always forgot how
to do the trick). \code{objList()} lists all objects, \code{funList()} just the
functions in a package.
}
\examples{

funList("bedrock")

}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) \emph{The
New S Language}. Wadsworth & Brooks/Cole.
}
\seealso{
\code{\link{ls}}, \code{\link{ls.str}}, \code{\link{lsf.str}}

Other pkg.introspection:
\code{\link[=funArgs]{funArgs()}},
\code{\link[=funCalls]{funCalls()}},
\code{\link[=funKeywords]{funKeywords()}},
\code{\link[=getRdLabels]{getRdLabels()}},
\code{\link[=inDots]{inDots()}},
\code{\link[=mergeArgs]{mergeArgs()}},
\code{\link[=strX]{strX()}}
}
\concept{data-inspection}
\concept{package-utilities}
\concept{pkg.introspection}
