% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCD-LCM.R
\name{GCD-LCM}
\alias{GCD-LCM}
\alias{GCD}
\alias{LCM}
\title{Greatest Common Divisor and Least Common Multiple}
\usage{
GCD(..., na.rm = FALSE)

LCM(..., na.rm = FALSE)
}
\arguments{
\item{...}{integer or logical vectors}

\item{na.rm}{logical. Should missing values (including NaN) be removed?}
}
\value{
A numeric (integer) value.
}
\description{
Calculates the greatest common divisor (GCD) and least common multiple (LCM)
of all the values present in its arguments.
}
\details{
The computation is based on the Euclidean algorithm without using the
extended version. The greatest common divisor for all numbers in the integer
vector \code{x} will be computed (the multiple GCD). Negative values are
allowed and enter via their absolute value; logical vectors are coerced
to integer.
}
\note{
The following relation is always true:

\code{n * m = GCD(n, m) * LCM(n, m)}
}
\examples{
GCD(12, 10)
GCD(144, 233)    # Fibonacci numbers are relatively prime to each other

LCM(12, 10)
LCM(144, 233)    # = 144 * 233

# all elements will be flattened by unlist
GCD(2, 3, c(5, 7) * 11)
GCD(c(2*3, 3*5, 5*7))
LCM(c(2, 3, 5, 7) * 11)
LCM(2*3, 3*5, 5*7)

}
\seealso{
Other number.theory:
\code{\link[=digitSum]{digitSum()}},
\code{\link[=divisors]{divisors()}},
\code{\link[=factorize]{factorize()}},
\code{\link[=fibonacci]{fibonacci()}},
\code{\link[=isOdd]{isOdd()}},
\code{\link[=isPrime]{isPrime()}},
\code{\link[=primes]{primes()}}
}
\concept{number-theory}
\concept{number.theory}
\concept{numerical-methods}
