% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDotsArg.R
\name{getDotsArg}
\alias{getDotsArg}
\title{Get a single argument from dots with default}
\usage{
getDotsArg(dots, name, default = NULL)
}
\arguments{
\item{dots}{Named list (usually list(...))}

\item{name}{Character string, argument name}

\item{default}{Default value if argument not present}
}
\value{
The value of the argument or default
}
\description{
Lightweight helper to extract a named argument from a list (typically list(...)).
}
