% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isEuclid.R
\name{isEuclid}
\alias{isEuclid}
\title{Test if a distance matrix is Euclidean}
\usage{
isEuclid(distmat, tol = 0.0000001)
}
\arguments{
\item{distmat}{An object of class \code{dist}.}

\item{tol}{Numeric tolerance for detecting negative eigenvalues.}
}
\value{
A logical scalar. Returns \code{TRUE} if the distance matrix is
(approximately) Euclidean, otherwise \code{FALSE}.
}
\description{
Checks whether a distance matrix corresponds to Euclidean distances.
}
\details{
The test is based on the eigenvalues of the double-centered squared
distance matrix \eqn{B = -\frac{1}{2} J D^2 J}. A distance matrix is
Euclidean if and only if \eqn{B} is positive semi-definite, i.e., all
eigenvalues are non-negative (within numerical tolerance).

The returned logical value carries additional diagnostic information
as attributes:
\itemize{
\item \code{eigenvalues}: Eigenvalues of the centered matrix
\item \code{min_eigenvalue}: Smallest eigenvalue
\item \code{tol}: Tolerance used for the test
}
}
\examples{
d <- dist(matrix(rnorm(20), ncol = 2))
res <- isEuclid(d)
res

# Access diagnostics
attr(res, "eigenvalues")
attr(res, "min_eigenvalue")

}
\concept{correlation}
\concept{data-inspection}
\concept{mathematics}
