% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isZero.R
\name{isZero}
\alias{isZero}
\title{Check a Vector For Being Zero}
\usage{
isZero(x, tol = sqrt(.Machine$double.eps), na.rm = FALSE)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values.}

\item{tol}{tolerance to be used}

\item{na.rm}{logical, indicating whether \code{NA} values should be stripped
before the computation proceeds. Defaults to \code{FALSE}.}
}
\value{
logical vector of the same dimension as x.
}
\description{
Test if x is zero. This is done by checking if the numeric value is
below the machine tolerance.
}
\examples{
# ... These are people who live in ignorance of the Floating Point Gods.
# These pagans expect ... (Burns, 2011)" the following to be TRUE:
(.1 - .3 / 3) == 0

# they might be helped by
isZero(.1 - .3 / 3)

}
\seealso{
\code{\link{is.integer}}
}
\author{
R-Core, Andri Signorell \href{mailto:andri@signorell.net}{andri@signorell.net}, Thomas W. Yee
}
\keyword{logic}
