% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelX.R
\name{Label}
\alias{Label}
\alias{Label<-}
\alias{Labels}
\alias{Labels<-}
\alias{Unit}
\alias{Unit<-}
\title{Label, Unit Attribute of an Object}
\usage{
Label(x)

Label(x) <- value

Labels(x) <- value

Labels(x)
}
\arguments{
\item{x}{any object}

\item{value}{a single string describing the object}
}
\value{
\code{Label} and \code{Unit} return the label attribute of x, if
any; otherwise, NULL.
}
\description{
Set and retrieve the \code{label}, resp. \code{unit} attribute of \code{x}.
This can be helpful for documenting the specific meaning of a variable, of
an entire data.frame or any other object. For single vectors it can be
useful to store the unit.
}
\details{
The label should consist of a single text (length of 1). The text may
contain line feeds. It can be deleted by setting the label to \code{NULL}.

\code{Labels()} can be used to retrieve and assign vectorized labels to
data.frames or lists.
}
\examples{

set.seed(123)
d.example <- data.frame(
  id        = 1:12,
  age       = round(rnorm(12, mean = 55, sd = 8)),
  bmi       = round(rnorm(12, mean = 27, sd = 4), 1),
  sex       = factor(rep(c("female", "male"), each = 6)),
  treatment = factor(rep(c("control", "drug"), times = 6))
  )

# add a descriptive label to a variable
Label(d.example$age) <- "Age in years (numeric)"

# technically just appending the text as attribute to the variable
attributes(d.example$age)

# label is supported while describing data
# Desc(d.example$age)

# The label can be deleted by setting it to NULL
Label(d.example$age) <- NULL

# Labelling the columns of a data.frame in one command
Labels(d.example) <- c("Age in years (numeric)",
                       "Body mass index (numeric)",
                       "Sex (factor with 2 levels: 'female', 'male')",
                       "Treatment group (factor with 2 levels: 'control', 'drug')"
                       )

}
\seealso{
A more elaborated label version can be found in package \pkg{Hmisc}
\code{\link[Hmisc]{label}()}.
}
\author{
Andri Signorell \href{mailto:andri@signorell.net}{andri@signorell.net}
}
\keyword{misc}
\keyword{utilities}
