% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\alias{logitInv}
\title{Generalized logit and Inverse logit Function}
\usage{
logit(x, min = 0, max = 1)

logitInv(x, min = 0, max = 1)
}
\arguments{
\item{x}{value(s) to be transformed}

\item{min}{lower end of logit interval}

\item{max}{upper end of logit interval}
}
\value{
Transformed value(s).
}
\description{
Compute generalized logit and generalized inverse logit functions.
}
\details{
The generalized logit function takes values on \verb{[min, max]} and transforms
them to span \eqn{[-\infty, \infty ]}{[-Inf, Inf]}. \cr It is defined as:

\deqn{y = log\left (\frac{p}{1-p} \right ) \;\;\; \; \textup{where} \; \;\;
p=\frac{x-min}{max-min}}{y = log(p/(1-p)) where p=(x-min)/(max-min)}

The generalized inverse logit function provides the inverse transformation:

\deqn{x = p' \cdot (max-min) + min \;\;\; \; \textup{where} \; \;\;
p'=\frac{exp(y)}{1+exp(y)}}{x = p' * (max-min) + min where p' =
exp(y)/(1+exp(y))}
}
\examples{


x <- seq(0,10, by=0.25)
xt <- logit(x, min=0, max=10)
cbind(x,xt)

y <- logitInv(xt, min=0, max=10)
cbind(x, xt, y)
}
\seealso{
\code{\link[car]{logit}}
}
\author{
Gregory R. Warnes \email{greg@warnes.net}
}
\keyword{math}
