% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveAvg.R
\name{moveAvg}
\alias{moveAvg}
\title{Moving Average}
\usage{
moveAvg(
  x,
  order,
  align = c("center", "left", "right"),
  endrule = c("NA", "keep", "constant")
)
}
\arguments{
\item{x}{univariate time series.}

\item{order}{order of moving average.}

\item{align}{specifies whether result should be centered (default),
left-aligned or right-aligned.}

\item{endrule}{Character string indicating how the values at the beginning
and the end of the data should be treated.
\describe{
\item{"keep"}{
Keeps the first and last \eqn{k_2}{k2} values at both ends,
where \eqn{k_2}{k2} is the half-bandwidth \code{k2 = k \%/\% 2},
i.e., \code{y[j] = x[j]} for
\eqn{j \in \{1,\ldots,k_2\}} and
\eqn{j \in \{n-k_2+1,\ldots,n\}}.
}
\item{"constant"}{
Fills the ends with the first and last calculated value in the output
array, e.g., \code{out[1:k2] = out[k2+1]}.
}
\item{"NA"}{
The default. Leaves the values as \code{NA},
as returned by \code{\link{filter}}.
}
}}
}
\value{
Returns a vector of the same size and same class as x.
}
\description{
Compute a simple moving average (running mean).
}
\details{
The implementation is using the function \code{filter} to calculate the
moving average.
}
\examples{

moveAvg(AirPassengers, order=5)

}
\seealso{
There's a faster implementation of running mean in the package
\pkg{caTools} \code{\link[caTools]{runmean}()} and a slower one in
\pkg{forecast} \code{\link[forecast]{ma}()}. There's similar code in
\code{\link[aurora]{midx}()}.
}
\keyword{univar}
