% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naIf.R
\name{naIf}
\alias{naIf}
\title{Replace Values with NA}
\usage{
naIf(x, values)
}
\arguments{
\item{x}{A vector.}

\item{values}{Values to be replaced by \code{NA}.}
}
\value{
A vector of the same type as \code{x}.
}
\description{
Replaces specified values in a vector with \code{NA}, in the manner of
SQL's \code{NULLIF}. This is the complementary operation to
\code{\link{coalesceX}}.
}
\examples{
naIf(c(1, 2, 99, 3, 99), 99)
naIf(c("a", "b", "n/a", ""), c("n/a", ""))

}
\seealso{
\code{\link{coalesceX}}

Other vector.ops:
\code{\link[=closest]{closest()}},
\code{\link[=coalesceX]{coalesceX()}},
\code{\link[=locf]{locf()}},
\code{\link[=midx]{midx()}},
\code{\link[=moveAvg]{moveAvg()}},
\code{\link[=naReplace]{naReplace()}},
\code{\link[=nz]{nz()}},
\code{\link[=pairApply]{pairApply()}},
\code{\link[=quot]{quot()}},
\code{\link[=setLength]{setLength()}},
\code{\link[=trim]{trim()}},
\code{\link[=vRot]{vRot()}},
\code{\link[=vShift]{vShift()}},
\code{\link[=winsorize]{winsorize()}}
}
\concept{missing-value}
\concept{vector.ops}
