% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naIf.R
\name{naIf}
\alias{naIf}
\title{Replace Values with NA}
\usage{
naIf(x, values)
}
\arguments{
\item{x}{A vector.}

\item{values}{Values to be replaced by \code{NA}.}
}
\value{
A vector of the same type as \code{x}.
}
\description{
Replaces specified values in a vector with \code{NA}.
}
\seealso{
Other vector.ops:
\code{\link[=Coalesce]{Coalesce()}},
\code{\link[=closest]{closest()}},
\code{\link[=midx]{midx()}},
\code{\link[=moveAvg]{moveAvg()}},
\code{\link[=naReplace]{naReplace()}},
\code{\link[=nz]{nz()}},
\code{\link[=pairApply]{pairApply()}},
\code{\link[=quot]{quot()}},
\code{\link[=rankX]{rankX()}},
\code{\link[=splitAt]{splitAt()}},
\code{\link[=trim]{trim()}},
\code{\link[=unwhich]{unwhich()}},
\code{\link[=vRot]{vRot()}},
\code{\link[=vShift]{vShift()}},
\code{\link[=winsorize]{winsorize()}}
}
\concept{data-manipulation}
\concept{missing-data}
\concept{vector-manipulation}
\concept{vector.ops}
