% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naIf.R
\name{naIf}
\alias{naIf}
\title{Replace Values with NA}
\usage{
naIf(x, values)
}
\arguments{
\item{x}{A vector.}

\item{values}{Values to be replaced by \code{NA}.}
}
\value{
A vector of the same type as \code{x}.
}
\description{
Replaces specified values in a vector with \code{NA}.
}
\seealso{
Other vector.ops: 
\code{\link{Coalesce}()},
\code{\link{closest}()},
\code{\link{midx}()},
\code{\link{moveAvg}()},
\code{\link{naReplace}()},
\code{\link{nz}()},
\code{\link{pairApply}()},
\code{\link{quot}()},
\code{\link{rankX}()},
\code{\link{splitAt}()},
\code{\link{trim}()},
\code{\link{unwhich}()},
\code{\link{vRot}()},
\code{\link{vShift}()},
\code{\link{winsorize}()}
}
\concept{data-manipulation}
\concept{missing-data}
\concept{vector-manipulation}
\concept{vector.ops}
