% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naIf.R
\name{naIf}
\alias{naIf}
\title{Replace Values with NA}
\usage{
naIf(x, values)
}
\arguments{
\item{x}{A vector.}

\item{values}{Values to be replaced by \code{NA}.}
}
\value{
A vector of the same type as \code{x}.
}
\description{
Replaces specified values in a vector with \code{NA}.
}
