% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naReplace.R
\name{naReplace}
\alias{naReplace}
\title{Replace Missing Values}
\usage{
naReplace(x, value)
}
\arguments{
\item{x}{A vector.}

\item{value}{Replacement value.}
}
\value{
A vector of the same type as \code{x}.
}
\description{
Replaces missing values (\code{NA}) in a vector with a specified value.
}
