% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric-conversions.R
\name{numeric-conversions}
\alias{numeric-conversions}
\alias{hexToDec}
\alias{binToDec}
\alias{decToBin}
\alias{octToDec}
\alias{decToOct}
\alias{decToHex}
\alias{romanToInt}
\title{Converts Numbers From Binmode, Octmode or Hexmode to Decimal and Vice Versa}
\usage{
hexToDec(x)

decToHex(x)

octToDec(x)

decToOct(x)

binToDec(x)

decToBin(x)

romanToInt(x)
}
\arguments{
\item{x}{a vector of numbers, resp. alphanumerical representation of numbers
(hex), to be converted.}
}
\value{
A numeric or character vector of the same length as x containing the
converted values. Binary, octal and decimal values are numeric, hex-values
are returned as class \code{hexmode}.
}
\description{
These functions convert numbers from one base to another. There are several
solutions for this problem out there, but the naming is quite heterogeneous
and so consistent function names might be helpful.
}
\details{
\code{binToDec()} converts numbers from binary mode into decimal values.
\code{decToBin} does it the other way round.\cr Oct means octal
system and hex hexadecimal.
}
\examples{

decToBin(c(17, 25))
binToDec(c(101, 11101))

decToOct(c(17, 25))
octToDec(c(11, 77))

decToHex(c(17, 25))
hexToDec(c("FF", "AA", "ABC"))

}
\seealso{
\code{\link{strtoi}}

Other number.theory:
\code{\link[=digitSum]{digitSum()}},
\code{\link[=factorize]{factorize()}},
\code{\link[=fibonacci]{fibonacci()}},
\code{\link[=isOdd]{isOdd()}},
\code{\link[=isPrime]{isPrime()}},
\code{\link[=primes]{primes()}}
}
\concept{data-manipulation}
\concept{number-theory}
\concept{number.theory}
\concept{string-manipulation}
\keyword{arith}
