% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nunique.R
\name{nUnique}
\alias{nUnique}
\title{Count Unique Values}
\usage{
nUnique(x, na.rm = FALSE)
}
\arguments{
\item{x}{a vector.}

\item{na.rm}{logical. Should missing values (\code{NA}) be removed before
counting unique values? Defaults to \code{FALSE}.}
}
\value{
an integer of length one.
}
\description{
Returns the number of unique vector elements.
}
\examples{
nUnique(c(1, 1, 2, 3))

nUnique(c(1, 1, 2, NA))

nUnique(c(1, 1, 2, NA), na.rm = TRUE)

}
\seealso{
\code{\link[=nlevels]{nlevels()}}, \code{\link[=isLowCardinality]{isLowCardinality()}} to check whether \code{x}
has at most a given number of unique values, without counting
all of them first.

Other data.predicate:
\code{\link[=flags]{flags()}},
\code{\link[=isDichotomous]{isDichotomous()}},
\code{\link[=isEuclid]{isEuclid()}},
\code{\link[=isLowCardinality]{isLowCardinality()}},
\code{\link[=isNumeric]{isNumeric()}},
\code{\link[=isWholeLike]{isWholeLike()}},
\code{\link[=isZero]{isZero()}}
}
\concept{cardinality}
\concept{data-inspection}
\concept{data.predicate}
