% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nunique.R
\name{nunique}
\alias{nunique}
\title{Count Unique Values}
\usage{
nunique(x, na.rm = FALSE)
}
\arguments{
\item{x}{A vector.}

\item{na.rm}{Logical. Should missing values (\code{NA}) be removed before
counting unique values? Defaults to \code{FALSE}.}
}
\value{
An integer of length one.
}
\description{
Returns the number of unique elements in a vector.
}
\examples{
nunique(c(1, 1, 2, 3))

nunique(c(1, 1, 2, NA))

nunique(c(1, 1, 2, NA), na.rm = TRUE)

}
