% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nUnique.R
\name{nUnique}
\alias{nUnique}
\title{Count Unique Values}
\usage{
nUnique(x, na.rm = FALSE)
}
\arguments{
\item{x}{A vector.}

\item{na.rm}{Logical. Should missing values (\code{NA}) be removed before
counting unique values? Defaults to \code{FALSE}.}
}
\value{
An integer of length one.
}
\description{
Returns the number of unique elements in a vector.
}
\examples{
nUnique(c(1, 1, 2, 3))

nUnique(c(1, 1, 2, NA))

nUnique(c(1, 1, 2, NA), na.rm = TRUE)

}
\seealso{
\code{\link[=isLowCardinality]{isLowCardinality()}} to check whether \code{x} has fewer than a given
number of unique values, without counting all of them first.
}
