% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz.R
\name{nz}
\alias{nz}
\title{Extract Non-Zero Values}
\usage{
nz(x, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{A numeric vector.}

\item{tol}{Tolerance passed to \code{\link{isZero}}.}
}
\value{
A vector containing only the non-zero elements of \code{x}.
}
\description{
Returns all non-zero elements of a vector. Zeroness is determined by
\code{\link{isZero}}, i.e. within a numerical tolerance.
}
\details{
\code{NA} elements are not considered zero and are retained in the
result.
}
\examples{
nz(c(0, 1, 2, 0, 3))
nz(c(1e-20, 1, NA))

}
\seealso{
\code{\link{isZero}}

Other vector.ops:
\code{\link[=closest]{closest()}},
\code{\link[=coalesceX]{coalesceX()}},
\code{\link[=locf]{locf()}},
\code{\link[=midx]{midx()}},
\code{\link[=moveAvg]{moveAvg()}},
\code{\link[=naIf]{naIf()}},
\code{\link[=naReplace]{naReplace()}},
\code{\link[=pairApply]{pairApply()}},
\code{\link[=quot]{quot()}},
\code{\link[=setLength]{setLength()}},
\code{\link[=trim]{trim()}},
\code{\link[=vRot]{vRot()}},
\code{\link[=vShift]{vShift()}},
\code{\link[=winsorize]{winsorize()}}
}
\concept{missing-value}
\concept{vector.ops}
