% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nz.R
\name{nz}
\alias{nz}
\title{Extract Non-Zero Values}
\usage{
nz(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A vector containing only the non-zero elements of \code{x}.
}
\description{
Returns all non-zero elements of a vector.
}
\details{
This is a convenience function primarily intended for use in
model formulas or quick vector filtering.
}
\seealso{
Other vector.ops:
\code{\link[=Coalesce]{Coalesce()}},
\code{\link[=closest]{closest()}},
\code{\link[=midx]{midx()}},
\code{\link[=moveAvg]{moveAvg()}},
\code{\link[=naIf]{naIf()}},
\code{\link[=naReplace]{naReplace()}},
\code{\link[=pairApply]{pairApply()}},
\code{\link[=quot]{quot()}},
\code{\link[=rankX]{rankX()}},
\code{\link[=splitAt]{splitAt()}},
\code{\link[=trim]{trim()}},
\code{\link[=unwhich]{unwhich()}},
\code{\link[=vRot]{vRot()}},
\code{\link[=vShift]{vShift()}},
\code{\link[=winsorize]{winsorize()}}
}
\concept{data-inspection}
\concept{vector-manipulation}
\concept{vector.ops}
