% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairApply.R
\name{pairApply}
\alias{pairApply}
\title{Pairwise Calculations}
\usage{
pairApply(x, FUN = NULL, ..., symmetric = FALSE)
}
\arguments{
\item{x}{a list, a data.frame or a matrix with columns to be processed
pairwise.}

\item{FUN}{a function to be calculated. It is assumed, that the first 2
arguments denominate x and y.}

\item{\dots}{the dots are passed to FUN.}

\item{symmetric}{logical. Does the function yield the same result for FUN(x,
y) and FUN(y, x)? \cr If \code{TRUE} just the lower triangular matrix is
calculated and transposed. Default is FALSE.}
}
\value{
a matrix with the results of FUN.
}
\description{
Implements a logic to run pairwise calculations on the columns of a
data.frame or a matrix.
}
\details{
This code is based on the logic of \code{cor()} and extended for asymmetric
functions.
}
\examples{

# build a dataset
set.seed(1)
d.sub <- transform(
  data.frame(
    X1 = rnorm(n <- 300), 
    X3 = rnorm(n)),
  X2 = 0.8*X1 + rnorm(n),
  X4 = 0.5*X3 + rnorm(n)
  )

pairApply(d.sub, FUN = cor, method="spearman")

# user defined functions are ok as well
pairApply(d.sub, 
  FUN = function(x,y) 
    wilcox.test(as.numeric(x), as.numeric(y))$p.value, symmetric=TRUE)

}
\seealso{
\code{\link{outer}}, \code{\link{combPairs}},
\code{\link{pairwise.table}}

Other vector.ops:
\code{\link[=Coalesce]{Coalesce()}},
\code{\link[=closest]{closest()}},
\code{\link[=midx]{midx()}},
\code{\link[=moveAvg]{moveAvg()}},
\code{\link[=naIf]{naIf()}},
\code{\link[=naReplace]{naReplace()}},
\code{\link[=nz]{nz()}},
\code{\link[=quot]{quot()}},
\code{\link[=rankX]{rankX()}},
\code{\link[=splitAt]{splitAt()}},
\code{\link[=trim]{trim()}},
\code{\link[=unwhich]{unwhich()}},
\code{\link[=vRot]{vRot()}},
\code{\link[=vShift]{vShift()}},
\code{\link[=winsorize]{winsorize()}}
}
\concept{combinatorics}
\concept{data-manipulation}
\concept{vector-manipulation}
\concept{vector.ops}
\keyword{manip}
