% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permn.R
\name{permn}
\alias{permn}
\title{Set of Permutations}
\usage{
permn(x, sortResults = FALSE)
}
\arguments{
\item{x}{Atomic vector. Missing values are not supported.}

\item{sortResults}{Logical scalar. If \code{TRUE}, the result matrix is sorted
using \code{\link[=sortX]{sortX()}}. Default is \code{FALSE}.}
}
\value{
A matrix containing all distinct permutations of \code{x}, one
permutation per row.
}
\description{
Returns all distinct permutations of a vector. Repeated values in \code{x}
are treated as indistinguishable, so duplicated permutations are not returned.
}
\examples{
permn(letters[2:5])
permn(2:5)

# repeated elements are handled as indistinguishable
permn(c("a", "b", "c", "a"))

}
\seealso{
\code{\link[utils:combn]{utils::combn()}}, \code{\link[base:factorial]{base::factorial()}}, \link{combPairs}

Other combinatorics:
\code{\link[=combN]{combN()}},
\code{\link[=combPairs]{combPairs()}},
\code{\link[=combSet]{combSet()}},
\code{\link[=randGroupSplit]{randGroupSplit()}},
\code{\link[=sampleX]{sampleX()}},
\code{\link[=unwhich]{unwhich()}}
}
\concept{combinatorics}
\concept{sampling}
