% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision.R
\name{precision}
\alias{precision}
\alias{nDec}
\alias{prec}
\alias{frac}
\alias{maxDigits}
\title{Precision, Decimal Places and Fractional Part of a Numeric Value}
\usage{
nDec(x)

prec(x)

frac(x, dpwr = NA)

maxDigits(x)
}
\arguments{
\item{x}{the numeric value (or a vector of numerics), whose fractional part
is to be calculated.}

\item{dpwr}{power of 10 for a factor z, the fractional part will be
multiplied with. The result will be returned rounded to integer. Defaults to
\code{NA} and will then be ignored.}
}
\value{
\itemize{
\item \code{nDec()}: an integer vector of the same length as \code{x};
\code{NA} elements yield \code{NA}.
\item \code{prec()}: a single numeric value, the finest precision found
across all (non-missing) elements of \code{x}. Returns 1 if all values
are zero and \code{NA} if no non-missing values are left.
\item \code{frac()}: a numeric vector of the same length as \code{x}.
\item \code{maxDigits()}: a single integer value.
}
}
\description{
\code{nDec()} returns the number of decimals.\cr
\code{prec()} returns the precision of a number x, i.e. the smallest positional
value of its last significant digit (e.g. 0.001 for 3.142).\cr
\code{frac()} returns the fractional part of a numeric value.\cr
\code{maxDigits()} returns the maximal number of decimal digits in \code{x}.
}
\examples{

x <- rnorm(5)*100
x
frac(x)

# multiply by 10^4
frac(x, dpwr=4)

maxDigits(c(1.25, 1.8, 12.0, 1.00000))

x <- c("0.0000", "0", "159.283", "1.45e+10", "1.4599E+10" )
nDec(x)
prec(as.numeric(x))

}
\seealso{
\code{\link{format.info}}, \code{\link{as.integer}},
\code{\link{trunc}}

Other math.utils:
\code{\link[=crossProd]{crossProd()}},
\code{\link[=crossProdN]{crossProdN()}},
\code{\link[=dotProd]{dotProd()}},
\code{\link[=isLowCardinality]{isLowCardinality()}},
\code{\link[=linScale]{linScale()}},
\code{\link[=logit]{logit()}},
\code{\link[=nUnique]{nUnique()}},
\code{\link[=percentRank]{percentRank()}},
\code{\link[=rankX]{rankX()}},
\code{\link[=roundTo]{roundTo()}},
\code{\link[=unirootAll]{unirootAll()}},
\code{\link[=untable]{untable()}}
}
\concept{math.utils}
\concept{number-formatting}
\keyword{arith}
