% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nDec.R
\name{precision}
\alias{precision}
\alias{nDec}
\alias{frac}
\alias{maxDigits}
\alias{ndec}
\alias{prec}
\title{Fractional Part and Maximal Digits of a Numeric Value}
\usage{
nDec(x)

prec(x)

frac(x, dpwr = NA)

maxDigits(x)
}
\arguments{
\item{x}{the numeric value (or a vector of numerics), whose fractional part
is to be calculated.}

\item{dpwr}{power of 10 for a factor z, the fractional part will be
multiplied with. The result will be returned rounded to integer. Defaults to
\code{NA} and will then be ignored.}
}
\description{
\code{frac()} returns the fractional part of a numeric value.
\code{maxDigits()} return the number of digits in \code{x}. \cr
\code{nDec()} returns the number of decimals.\cr \code{prec()} returns the
precision of a number \code{x}.
}
\examples{

x <- rnorm(5)*100
x
frac(x)

# multiply by 10^4
frac(x, dpwr=4)

maxDigits(c(1.25, 1.8, 12.0, 1.00000))

x <- c("0.0000", "0", "159.283", "1.45e+10", "1.4599E+10" )
nDec(x)
prec(as.numeric(x))

}
\seealso{
\code{\link{format.info}}, \code{\link{as.integer}},
\code{\link{trunc}}
}
\author{
Andri Signorell \href{mailto:andri@signorell.net}{andri@signorell.net}
}
\keyword{arith}
