% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quot.R
\name{quot}
\alias{quot}
\title{Lagged Quotients}
\usage{
quot(x, lag = 1L, quotients = 1L, ...)
}
\arguments{
\item{x}{a numeric vector or matrix containing the values to be used for
calculating the quotients.}

\item{lag}{an integer indicating which lag to use.}

\item{quotients}{an integer indicating the order of the quotient.}

\item{\dots}{further arguments to be passed to or from methods.}
}
\value{
If \code{x} is a vector of length \code{n} and \code{quotients = 1},
then the computed result is equal to the successive quotients
\code{x[(1+lag):n] - x[1:(n-lag)]}.

If \code{quotients} is larger than one this algorithm is applied recursively
to \code{x}.  Note that the returned value is a vector which is shorter than
\code{x}.

If \code{x} is a matrix then the division operations are carried out on each
column separately.
}
\description{
Returns suitably lagged and iterated quotients
}
\details{
\code{\link{NA}}'s propagate.
}
\examples{

quot(1:10, 2)
quot(1:10, 2, 2)
x <- cumprod(cumprod(1:10))
quot(x, lag = 2)
quot(x, quotients = 2)

}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) \emph{The
New S Language}.  Wadsworth & Brooks/Cole.
}
\seealso{
\code{\link{diff}}
}
\keyword{arith}
