% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randGroupSplit.R
\name{randGroupSplit}
\alias{randGroupSplit}
\title{Randomly Split a Vector into Groups of Given Sizes}
\usage{
randGroupSplit(x, groupSizes)
}
\arguments{
\item{x}{A vector containing the elements to be split into groups.}

\item{groupSizes}{An integer vector specifying the sizes of the groups.
The sum of \code{groupSizes} must equal \code{length(x)}.}
}
\value{
A list of vectors, where each element corresponds to one group.
The length of the list equals \code{length(groupSizes)}.
}
\description{
Randomly assigns the elements of a vector \code{x} into groups with
predefined sizes given by \code{groupSizes}. The grouping is performed
without replacement and each element is assigned to exactly one group.
}
\details{
This function is useful for random group assignments, for example in
teaching settings, simulations, or experimental designs where groups
of unequal sizes are required.
}
\examples{
# Split letters into 3 groups of sizes 4, 3, and 5
set.seed(123)
randGroupSplit(LETTERS[1:12], groupSizes = c(4, 3, 5))

}
\seealso{
Other combinatorics:
\code{\link[=combN]{combN()}},
\code{\link[=combPairs]{combPairs()}},
\code{\link[=combSet]{combSet()}},
\code{\link[=permn]{permn()}},
\code{\link[=sampleX]{sampleX()}}
}
\concept{combinatorics}
\concept{data-manipulation}
\concept{vector-manipulation}
