% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range-operators.R
\name{range-operators}
\alias{range-operators}
\alias{\%:\%}
\alias{\%::\%}
\title{Select a Range Between Two Elements (Lazy)}
\usage{
x \%:\% rng

x \%::\% rng
}
\arguments{
\item{x}{a vector.}

\item{rng}{a vector of length 2: \code{c(from, to)}. May contain \code{NA} to match
missing values in \code{x}.}
}
\value{
a subset of \code{x}, from the first occurrence of \code{rng[1]} to the
first occurrence of \code{rng[2]}.

a subset of \code{x}, from the first occurrence of \code{rng[1]} to the
last occurrence of \code{rng[2]}.
}
\description{
Returns the elements of \code{x} from the first occurrence of \code{rng[1]} to the
first occurrence of \code{rng[2]}. Equivalent to a non-greedy ("lazy") match,
analogous to lazy quantifiers in regular expressions.

Returns the elements of \code{x} from the first occurrence of \code{rng[1]} to the
\emph{last} occurrence of \code{rng[2]}. Equivalent to a greedy match, analogous to
greedy quantifiers in regular expressions.
}
\examples{
letters \%:\% c("c", "g")

x <- c("a", "b", "c", "d", "c", "e", "f", "c")
x \%:\% c("c", "e")

colnames(Pizza) \%:\% c("price", "nps")

letters \%::\% c("c", "g")

x <- c("a", "b", "c", "d", "c", "e", "f", "c")
x \%::\% c("c", "e")

}
\seealso{
the \verb{\%::\%} operator for the greedy variant, which uses the
\emph{last} occurrence of \code{rng[2]}.

the \verb{\%::\%}. operator for the lazy variant, which uses the
\emph{first} occurrence of \code{rng[2]}.
}
