% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdTitle.R
\name{rdTitle}
\alias{rdTitle}
\title{Extract the Title from an Rd Help File}
\usage{
rdTitle(topic, man = "man")
}
\arguments{
\item{topic}{a single character string giving the topic (function name or
alias) to look up.}

\item{man}{a single character string giving the path to the directory
containing \file{.Rd} files.  Defaults to \code{"man"}, i.e. the
\file{man/} subdirectory of the current working directory.}
}
\value{
a single character string with the title, trimmed of leading and
trailing whitespace.  Stops with an error if \code{topic} is not found.
}
\description{
Searches all \file{.Rd} files in a package's \file{man/} directory for a
given topic (matched against \verb{\\\\alias} entries) and returns its
\verb{\\\\title} string.
}
\examples{
# a minimal man/ directory to search in
man <- file.path(tempdir(), "man")
dir.create(man, showWarnings = FALSE)

writeLines(c("\\\\\\\\name{foo}", "\\\\\\\\alias{foo}", "\\\\\\\\alias{bar}",
             "\\\\\\\\title{A Minimal Help Page}",
             "\\\\\\\\description{Nothing to see here.}"),
           file.path(man, "foo.Rd"))

rdTitle("foo", man = man)
rdTitle("bar", man = man)          # aliases are matched as well
rdTitle("nothing", man = man)      # NA

unlink(man, recursive = TRUE)

}
\seealso{
\code{\link[tools:parse_Rd]{tools::parse_Rd()}}

Other pkg.funinfo:
\code{\link[=funArgs]{funArgs()}},
\code{\link[=funCalls]{funCalls()}},
\code{\link[=funKeywords]{funKeywords()}},
\code{\link[=funList]{funList()}},
\code{\link[=rdLabels]{rdLabels()}}
}
\concept{introspection}
\concept{pkg.funinfo}
\concept{programming}
