% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdTitle.R
\name{rdTitle}
\alias{rdTitle}
\title{Extract the Title from an Rd Help File}
\usage{
rdTitle(topic, man = "man")
}
\arguments{
\item{topic}{a single character string giving the topic (function name or
alias) to look up.}

\item{man}{a single character string giving the path to the directory
containing \file{.Rd} files.  Defaults to \code{"man"}, i.e. the
\file{man/} subdirectory of the current working directory.}
}
\value{
a single character string with the title, trimmed of leading and
trailing whitespace.  Stops with an error if \code{topic} is not found.
}
\description{
Searches all \file{.Rd} files in a package's \file{man/} directory for a
given topic (matched against \code{\\alias} entries) and returns its
\code{\\title} string.
}
\examples{
\dontrun{
rdTitle("mean")
rdTitle("lm", man = "path/to/pkg/man")
}

}
\seealso{
\code{\link[tools]{parse_Rd}}

Other pkg.funinfo:
\code{\link[=funArgs]{funArgs()}},
\code{\link[=funCalls]{funCalls()}},
\code{\link[=funKeywords]{funKeywords()}},
\code{\link[=funList]{funList()}},
\code{\link[=rdLabels]{rdLabels()}}
}
\concept{introspection}
\concept{pkg.funinfo}
\concept{programming}
