% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recycle.R
\name{recycle}
\alias{recycle}
\title{Recycle a List of Elements}
\usage{
recycle(..., maxdim = NULL, strict = FALSE)
}
\arguments{
\item{\dots}{a number of vectors of elements.}

\item{maxdim}{defines the maximal dimension, if set to \code{NULL} (default)
the maximal dimension of the list.}

\item{strict}{logical, if \code{TRUE} each element must have length 1 or
\code{maxdim}, so that no partial recycling (or truncation) can occur.
Default is \code{FALSE}.}
}
\value{
a list of the supplied elements\cr \code{attr(,"maxdim")} contains
the maximal dimension of the recycled list.
}
\description{
This function recycles all supplied elements to the maximal dimension.
}
\details{
If \code{maxdim} is smaller than the length of an element, that element
is truncated to the first \code{maxdim} values. Zero-length elements are
recycled to \code{NA} vectors of length \code{maxdim}. Both situations
are rejected when \code{strict = TRUE}.
}
\examples{

recycle(x=1:5, y=1, s=letters[1:2])

z <- recycle(x=letters[1:5], n=2:3, sep=c("-"," "))
sapply(1:attr(z, "maxdim"), function(i) paste(rep(z$x[i], times=z$n[i]),
                                        collapse=z$sep[i]))

}
\seealso{
\code{\link[=rep]{rep()}}, \code{\link[=replicate]{replicate()}}

Other pkg.args:
\code{\link[=callIf]{callIf()}},
\code{\link[=extractArgs]{extractArgs()}},
\code{\link[=getDotsArg]{getDotsArg()}},
\code{\link[=mergeArgs]{mergeArgs()}}
}
\concept{introspection}
\concept{pkg.args}
\concept{programming}
