% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recycle.R
\name{recycle}
\alias{recycle}
\title{Recyle a List of Elements}
\usage{
recycle(..., maxdim = NULL, strict = FALSE)
}
\arguments{
\item{\dots}{a number of vectors of elements.}

\item{maxdim}{defines the maximal dimension, if set to \code{NULL} (default)
the maximal dimension of the list.}

\item{strict}{defines if number of arguments must be 1 or maxdim.}
}
\value{
a list of the supplied elements\cr \code{attr(,"maxdim")} contains
the maximal dimension of the recyled list
}
\description{
This function recycles all supplied elments to the maximal dimension.
}
\examples{

recycle(x=1:5, y=1, s=letters[1:2])

z <- recycle(x=letters[1:5], n=2:3, sep=c("-"," "))
sapply(1:attr(z, "maxdim"), function(i) paste(rep(z$x[i], times=z$n[i]), 
                                        collapse=z$sep[i]))

}
\seealso{
\code{\link{rep}}, \code{\link{replicate}}

Other data.manipulation:
\code{\link[=appendRowNames]{appendRowNames()}},
\code{\link[=appendX]{appendX()}},
\code{\link[=as.array.xtabs]{as.array.xtabs()}},
\code{\link[=columnWrap]{columnWrap()}},
\code{\link[=combLevels]{combLevels()}},
\code{\link[=nf]{nf()}},
\code{\link[=parseSASDatalines]{parseSASDatalines()}},
\code{\link[=recodeX]{recodeX()}},
\code{\link[=revCode]{revCode()}},
\code{\link[=revX]{revX()}},
\code{\link{set-and-remove-attributes}},
\code{\link[=setNamesX]{setNamesX()}},
\code{\link[=sortX]{sortX()}},
\code{\link[=splitX]{splitX()}},
\code{\link[=stringsAsFactors]{stringsAsFactors()}},
\code{\link[=toBaseR]{toBaseR()}},
\code{\link{tolong_towide}},
\code{\link[=untable]{untable()}}
}
\concept{data-manipulation}
\concept{data.manipulation}
\concept{vector-manipulation}
\keyword{utilities}
