% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recycle.R
\name{recycle}
\alias{recycle}
\title{Recyle a List of Elements}
\usage{
recycle(..., strict = FALSE)
}
\arguments{
\item{\dots}{a number of vectors of elements.}

\item{strict}{defines if number of arguments must be 1 or maxdim.}
}
\value{
a list of the supplied elements\cr \code{attr(,"maxdim")} contains
the maximal dimension of the recyled list
}
\description{
This function recycles all supplied elments to the maximal dimension.
}
\examples{

recycle(x=1:5, y=1, s=letters[1:2])

z <- recycle(x=letters[1:5], n=2:3, sep=c("-"," "))
sapply(1:attr(z, "maxdim"), function(i) paste(rep(z$x[i], times=z$n[i]), 
                                        collapse=z$sep[i]))

}
\seealso{
\code{\link{rep}}, \code{\link{replicate}}
}
\keyword{utilities}
