% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolveGroups.R
\name{resolveGroups}
\alias{resolveGroups}
\title{Resolve Grouped Data}
\usage{
resolveGroups(x, g)
}
\arguments{
\item{x}{a numeric vector of observations, or a list of numeric vectors.}

\item{g}{a grouping variable of the same length as \code{x}.
Ignored when \code{x} is a list.}
}
\value{
A list containing:
\describe{
\item{x}{numeric response vector}
\item{g}{grouping factor}
\item{n}{total number of observations}
\item{k}{number of groups}
\item{group.sizes}{group sample sizes}
\item{group.names}{group labels}
\item{data.name}{character description of the input}
}
}
\description{
Standardizes grouped data supplied either as a numeric vector with
grouping variable or as a list of group-specific vectors.
}
\details{
Missing observations are removed and grouping information is returned
in a consistent format suitable for hypothesis tests, summaries,
effect-size calculations and plotting functions.
}
\seealso{
Other data.utils:
\code{\link[=resolveContingency]{resolveContingency()}}
}
\concept{data-manipulation}
\concept{data.utils}
\concept{grouped-data}
