% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revX.R
\name{revX}
\alias{revX}
\alias{revX.default}
\alias{revX.matrix}
\alias{revX.table}
\alias{revX.array}
\alias{revX.data.frame}
\title{Reverse Elements of a Vector, a Matrix, a Table, an Array or a Data.frame}
\usage{
revX(x, ...)

\method{revX}{default}(x, ...)

\method{revX}{array}(x, margin = seq_along(dim(x)), ...)

\method{revX}{matrix}(x, margin = seq_along(dim(x)), ...)

\method{revX}{table}(x, margin = seq_along(dim(x)), ...)

\method{revX}{data.frame}(x, margin = 1:2, ...)
}
\arguments{
\item{x}{a vector, a matrix or a higher dimensional table to be reversed.}

\item{\dots}{the dots are passed to the array interface.}

\item{margin}{vector of dimensions which to be reversed (1 for rows, 2 for
columns, etc.). If not defined, all dimensions will be reverted.}
}
\value{
an object of the same class and dimensions as \code{x}, with the
order of elements along \code{margin} reversed.
}
\description{
\code{revX} provides a reversed version of its argument. Unlike the basic
function, it does in higher-dimensional structures such as matrices not
reverse the elements, but the order of the rows and/or columns. It further
offers additional interfaces for higher dimensional arrays or tables.
}
\examples{

tab <- matrix(c(1, 11, 111,
                2, 22, 222,
                3, 33, 333), 
              byrow=TRUE, nrow=3,
              dimnames=list(mar1=1:3, mar2=c("a","b","c")))

revX(tab, margin=1)
revX(tab, margin=2)

# reverse both dimensions
revX(tab, margin=c(1, 2))

t(tab)

# reverse 3dimensional array
aa <- array(c(tab, 2 * tab), dim = c(3, 3, 2),
            dimnames = c(dimnames(tab), list(mar3 = c("A", "Z"))))

# reverse rows
revX(aa, 1)
# reverse columns
revX(aa, 2)
# reverse 3th dimension
revX(aa, 3)

# reverse all dimensions
revX(aa)
# same as
revX(aa, margin=(1:3))

}
\seealso{
\code{\link{rev}}, \code{\link{order}}, \code{\link{sort}},
\code{\link{seq}}

Other data.order:
\code{\link[=binaryTree]{binaryTree()}},
\code{\link[=sortX]{sortX()}}
}
\concept{data.order}
\concept{ordering}
\concept{reshape}
\keyword{manip}
