% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundTo.R
\name{roundTo}
\alias{roundTo}
\title{Round to Multiple}
\usage{
roundTo(x, multiple = 1, FUN = round)
}
\arguments{
\item{x}{numeric. The value to round.}

\item{multiple}{numeric. The multiple to which the number is to be rounded.
Default is 1.}

\item{FUN}{the rounding function as character or as expression. Can be one
out of \code{\link{trunc}}, \code{ceiling}, \code{round} (default) or
\code{floor}.}
}
\value{
the rounded value
}
\description{
Returns a number rounded to the nearest specified multiple.
}
\details{
There are several functions to convert to integers. \code{\link{round}}
rounds to the nearest integer or to any number of digits. Using a negative
number rounds to a power of ten, so that \code{round (x, -3)} rounds to
thousands. Each of \code{\link{trunc}}, \code{\link{floor}} and
\code{\link{ceiling}} round in a fixed direction, towards zero, down and up
respectively. \code{\link{round}} is documented to round to even, so
\code{round(2.5)} is \code{2}.

\code{\link{roundTo}} uses \code{round(x/multiple)*multiple} to get the
result. So if \code{x} is equally close to two multiples, the multiple with
the smaller absolute value will be returned when \code{round(x/multiple)} is
even (and the greater when it's odd).\cr If \code{FUN} is set to
\code{ceiling} it will always round up, and if set to \code{floor} it will
always round down. See examples for comparison).
}
\examples{

roundTo(10, 3)     # Rounds 10 to a nearest multiple of 3 (9)
roundTo(-10, -3)   # Rounds -10 to a nearest multiple of -3 (-9)

roundTo(1.3, 0.2)  # Rounds 1.3 to a nearest multiple of 0.2 (1.2)
roundTo(-1.3, 0.2) # Rounds -1.3 to a nearest multiple of 0.2 (-1.2)
roundTo(5, -2)     # Returns an error, because -2 and 5 have different signs

# Round down
roundTo(c(1,-1) * 1.2335, 0.05, floor)
roundTo(c(1,-1) * 1233.5, 100, floor)

# Round up
roundTo(c(1,-1) * 1.2335, 0.05, ceiling)
roundTo(c(1,-1) * 1233.5, 100, ceiling)

# Round towards zero
roundTo(c(1,-1) * 1.2335, 0.05, trunc)
roundTo(c(1,-1) * 1233.5, 100, trunc)


x <- c(-1.5,-1.3, 1.3, 1.5)
cbind(x =       x,
      round =   roundTo(x, 0.2, FUN=round),
      trunc =   roundTo(x, 0.2, FUN=trunc),
      ceiling = roundTo(x, 0.2, FUN=ceiling),
      floor =   roundTo(x, 0.2, FUN=floor)
)

x <- -10:10
cbind(x =       x,
      round =   roundTo(x, 2, FUN=round),
      trunc =   roundTo(x, 2, FUN=trunc),
      ceiling = roundTo(x, 2, FUN=ceiling),
      floor =   roundTo(x, 2, FUN=floor)
)


}
\seealso{
\code{\link{round}}, \code{\link{trunc}}, \code{\link{ceiling}},
\code{\link{floor}}

Other math.utils: 
\code{\link{crossProd}()},
\code{\link{crossProdN}()},
\code{\link{dotProd}()},
\code{\link{percentRank}()},
\code{\link{precision}},
\code{\link{ptInPoly}()},
\code{\link{unirootAll}()}
}
\concept{data-manipulation}
\concept{math.utils}
\concept{mathematics}
\keyword{manip}
