% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleX.R
\name{sampleX}
\alias{sampleX}
\alias{sampleX.data.frame}
\alias{sampleX.default}
\title{Random Samples and Permutations}
\usage{
sampleX(x, size, replace = FALSE, prob = NULL)

\method{sampleX}{data.frame}(x, size, replace = FALSE, prob = NULL)

\method{sampleX}{default}(x, size, replace = FALSE, prob = NULL)
}
\arguments{
\item{x}{either a vector of one or more elements from which to choose, or a
positive integer.}

\item{size}{a positive number, the number of items to choose from.}

\item{replace}{a non-negative integer giving the number of items to choose.}

\item{prob}{should sampling be with replacement?}
}
\value{
sampled elements in the same structure as x
}
\description{
\code{Sample} takes a sample of the specified size from the elements of x
using either with or without replacement. The function does the same as the
base::sample() and offers additionally an interface for data frames.
}
\examples{

sampleX(d.pizza, size=5)

}
\seealso{
\code{\link{sample}}
}
\author{
Andri Signorell \href{mailto:andri@signorell.net}{andri@signorell.net}
}
\keyword{IO}
\keyword{distribution}
