% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-remove-keep-attr.R
\name{setAttr-removeAttr-keepAttr}
\alias{setAttr-removeAttr-keepAttr}
\alias{setAttr}
\alias{removeAttr}
\alias{keepAttr}
\title{Set and Remove Object Attributes}
\usage{
setAttr(x, attrNames, attrValues)

removeAttr(x, attrNames = NULL)

keepAttr(x, attrNames)
}
\arguments{
\item{x}{Object to modify}

\item{attrNames}{Character vector of attribute names}

\item{attrValues}{Values for the attributes (only for setting)}
}
\value{
Modified object
}
\description{
Convenience helpers to add, remove, or selectively retain attributes
of an object.
}
\examples{
x <- runif(10)

x <- setAttr(
  x,
  attrNames = c("some_attr", "other_attr"),
  attrValues = c("First attribute", "Second attribute")
)

# remove single attribute
removeAttr(x, "other_attr")

# remove all attributes
removeAttr(x)

# keep only selected attributes, remove all others
r.lm <- lm(Fertility ~ ., swiss)
keepAttr(r.lm$terms, "class")
}
\seealso{
\code{\link{setNamesX}}, \code{\link{unname}}

Other data.manipulation:
\code{\link[=appendEnum]{appendEnum()}},
\code{\link[=appendRowNames]{appendRowNames()}},
\code{\link[=appendX]{appendX()}},
\code{\link[=as.array.xtabs]{as.array.xtabs()}},
\code{\link[=asBinary]{asBinary()}},
\code{\link[=asCDateFmt]{asCDateFmt()}},
\code{\link[=columnWrap]{columnWrap()}},
\code{\link[=combLevels]{combLevels()}},
\code{\link[=compareDataFrames]{compareDataFrames()}},
\code{\link[=dummy]{dummy()}},
\code{\link[=nf]{nf()}},
\code{\link[=recodeX]{recodeX()}},
\code{\link[=recycle]{recycle()}},
\code{\link[=renameX]{renameX()}},
\code{\link[=revCode]{revCode()}},
\code{\link[=revX]{revX()}},
\code{\link[=setNamesX]{setNamesX()}},
\code{\link[=sortX]{sortX()}},
\code{\link[=splitAt]{splitAt()}},
\code{\link[=splitX]{splitX()}},
\code{\link[=stringsAsFactors]{stringsAsFactors()}},
\code{\link[=toBaseR]{toBaseR()}}
}
\concept{data-manipulation}
\concept{data-structures}
\concept{data.manipulation}
