% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortX.R
\name{sortX}
\alias{sortX}
\alias{sortX.default}
\alias{sortX.data.frame}
\alias{sortX.matrix}
\alias{sortX.table}
\title{Sort a Vector, a Matrix, a Table or a Data.frame}
\usage{
sortX(x, ...)

\method{sortX}{default}(x, ...)

\method{sortX}{data.frame}(
  x,
  ord = NULL,
  decreasing = FALSE,
  factorsAsCharacter = TRUE,
  na.last = TRUE,
  ...
)

\method{sortX}{matrix}(x, ord = NULL, decreasing = FALSE, na.last = TRUE, ...)

\method{sortX}{table}(x, ord = NULL, decreasing = FALSE, na.last = TRUE, ...)
}
\arguments{
\item{x}{a numeric, complex. character or logical vector, a factor, a table
or a data.frame to be sorted.}

\item{\dots}{further arguments to be passed to or from methods.}

\item{ord}{vector of integers or columnames. Defines the columns in a table,
in a matrix or in a data.frame to be sorted for. \cr 0 means row.names, 1:n
the columns and n+1 the marginal sum. See examples.}

\item{decreasing}{logical. Should the sort be increasing or decreasing?}

\item{factorsAsCharacter}{logical. Should factors be sorted by the
alphabetic order of their labels or by the order or their levels.  Default
is \code{TRUE} (by labels).}

\item{na.last}{for controlling the treatment of \code{NAs}. If \code{TRUE},
missing values in the data are put last; if \code{FALSE}, they are put
first; if \code{NA}, they are removed (see \code{\link{order}}.)}
}
\value{
the sorted object.
}
\description{
Sort a vector, a matrix, a table or a data.frame. The base sort function
does not have an interface for classes other than vectors and coerces the
whole world to a vector. This means you get a sorted vector as result while
passing a matrix to \code{sort}.\cr \code{sortX} wraps the base sort function
and adds an interface for sorting the rows of the named 2-dimensional data
structures by the order of one or more of its columns.
}
\details{
The sort order for factors is the order of their levels (which is
particularly appropriate for ordered factors), and usually confusing for
unordered factors, whose levels may be defined in the sequence in which they
appear in the data (which normally is unordered).
}
\examples{
set.seed(3)
d.frm <- iris[sample(nrow(iris), 10), 
              c("Species", "Sepal.Length", "Sepal.Width")]

sortX(d.frm[,1])
# sortX follows the levels by default
levels(d.frm[,1])

sortX(x=d.frm, ord="Species", decreasing=FALSE)
# set factorsAsCharacter = TRUE, if alphabetical order is required
sortX(x=d.frm, ord="Species", decreasing=FALSE, factorsAsCharacter=TRUE)

sortX(x=d.frm, ord=c("Species","Sepal.Length"), factorsAsCharacter = TRUE)
sortX(x=d.frm, ord=c("Species","Sepal.Length"), factorsAsCharacter = FALSE)

sortX(x=d.frm, ord=c("Species","Sepal.Length"), decreasing=c(FALSE, TRUE),
  factorsAsCharacter = FALSE)

# Sorting tables
tab <- HairEyeColor[,,1]

sortX(x=tab, ord=c(0,2), decreasing=c(TRUE, FALSE))
sortX(x=tab, ord=2, decreasing=TRUE)

# partial matching ok:
sortX(tab, o=1, d=TRUE)

}
\seealso{
\code{\link{sort}}, \code{\link{order}}
}
