% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitAt.R
\name{splitAt}
\alias{splitAt}
\title{Split a Vector at Given Positions}
\usage{
splitAt(x, pos)
}
\arguments{
\item{x}{A vector to be split.}

\item{pos}{An integer vector of positions at which to split \code{x}.
Positions refer to indices in \code{x} where a new segment should start.}
}
\value{
A list of vectors, each representing a segment of \code{x}.
}
\description{
Splits a vector into consecutive segments at specified positions.
}
\details{
The function splits \code{x} into consecutive chunks defined by \code{pos}.
Internally, positions are sorted, duplicates are removed, and values outside
the valid index range \code{[1, length(x)]} are ignored.

Each element of the returned list corresponds to a contiguous subset of
\code{x}. The first segment always starts at position 1.
}
\examples{
x <- 1:10

# split at positions 4 and 7
splitAt(x, c(4, 7))

# unsorted and duplicate positions are handled
splitAt(x, c(7, 4, 4, 20))

}
\seealso{
\code{\link{split}}

Other vector.ops: 
\code{\link{Coalesce}()},
\code{\link{closest}()},
\code{\link{midx}()},
\code{\link{moveAvg}()},
\code{\link{naIf}()},
\code{\link{naReplace}()},
\code{\link{nz}()},
\code{\link{pairApply}()},
\code{\link{quot}()},
\code{\link{rankX}()},
\code{\link{trim}()},
\code{\link{unwhich}()},
\code{\link{vRot}()},
\code{\link{vShift}()},
\code{\link{winsorize}()}
}
\concept{data-manipulation}
\concept{vector-manipulation}
\concept{vector.ops}
