% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringsAsFactors.R
\name{stringsAsFactors}
\alias{stringsAsFactors}
\title{Convert Character Columns to Factors}
\usage{
stringsAsFactors(x, columns = NULL)
}
\arguments{
\item{x}{the data.frame.}

\item{columns}{names or indexes of the columns to be converted;
negative values can be used to omit columns.}
}
\value{
the given data.frame including the converted factors.
}
\description{
Strings as factors have recently been downgraded in base R's
\code{\link{data.frame}()} function. However,
it is still usually a good idea to encode string variables
as factors. This function helps to convert some or
all columns of a data.frame to factors.
}
\examples{
d.dat <- data.frame(char_x = LETTERS[1:5],
                    char_y = LETTERS[6:10],
                    n = 1:5)

# all character columns
str(stringsAsFactors(d.dat))
# only char_y
str(stringsAsFactors(d.dat, columns = "char_y"))
# only char_x
str(stringsAsFactors(d.dat, columns = "char_x"))
# all character columns except the second one ("char_y")
str(stringsAsFactors(d.dat, columns = -2))

}
\seealso{
Other data.recode:
\code{\link[=asBinary]{asBinary()}},
\code{\link[=combLevels]{combLevels()}},
\code{\link[=dummy]{dummy()}},
\code{\link[=mReplace]{mReplace()}},
\code{\link[=nf]{nf()}},
\code{\link[=recodeX]{recodeX()}},
\code{\link[=revCode]{revCode()}}
}
\concept{categorization}
\concept{data.recode}
\concept{type-coercion}
