% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringAsFactors.R
\name{stringsAsFactors}
\alias{stringsAsFactors}
\title{Convert Character Columns to Factors}
\usage{
stringsAsFactors(x, columns = NULL)
}
\arguments{
\item{x}{the data.frame}

\item{columns}{names or indexes of the columns to be converted;
negative values can be used to omit columns.}
}
\value{
the given data.frame including the converted factors
}
\description{
Strings as factors have recently been downgraded in base R's
\code{\link{data.frame}()} function. However,
it is still usually a good idea to encode string variables
as factors. This function helps to convert some or
all columns of a data.frame to factors.
}
\examples{
# get some data
d.dat <- data.frame(char_x=LETTERS[1:5],
                    char_y=LETTERS[6:10],
                    n=1:5)

# all characters
str(stringsAsFactors(d.dat))
# only char_y
str(stringsAsFactors(d.dat, columns="char_y"))
# only char_x
str(stringsAsFactors(d.dat, columns="char_x"))

# all characters, besides 2 (second column, so "char_y")
str(stringsAsFactors(d.dat, columns=-2))

}
\seealso{
Other data.manipulation:
\code{\link[=appendRowNames]{appendRowNames()}},
\code{\link[=appendX]{appendX()}},
\code{\link[=as.array.xtabs]{as.array.xtabs()}},
\code{\link[=columnWrap]{columnWrap()}},
\code{\link[=combLevels]{combLevels()}},
\code{\link[=nf]{nf()}},
\code{\link[=parseSASDatalines]{parseSASDatalines()}},
\code{\link[=recodeX]{recodeX()}},
\code{\link[=recycle]{recycle()}},
\code{\link[=revCode]{revCode()}},
\code{\link[=revX]{revX()}},
\code{\link{set-and-remove-attributes}},
\code{\link[=setNamesX]{setNamesX()}},
\code{\link[=sortX]{sortX()}},
\code{\link[=splitX]{splitX()}},
\code{\link[=toBaseR]{toBaseR()}},
\code{\link{tolong_towide}},
\code{\link[=untable]{untable()}}
}
\concept{data-manipulation}
\concept{data.manipulation}
\concept{factor-handling}
