% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strX.R
\name{strX}
\alias{strX}
\title{Extended str() with numbered variables}
\usage{
strX(object, ..., enumerate = TRUE, recursive = FALSE, strict.width = "cut")
}
\arguments{
\item{object}{Any R object.}

\item{...}{Additional arguments passed to \code{\link[utils:str]{utils::str()}}.}

\item{enumerate}{Logical. Should variables/elements be numbered?
Default is \code{TRUE}.}

\item{recursive}{Logical. Should nested list elements also be numbered?
Default is \code{FALSE}.}

\item{strict.width}{Character string passed to \code{\link[utils:str]{utils::str()}}.
Default is \code{"cut"}.}
}
\value{
Invisibly returns the character vector produced by \code{\link[utils:str]{utils::str()}}.
}
\description{
Wrapper around \code{\link[utils:str]{utils::str()}} that optionally numbers variables in
lists and data frames. Useful for large objects where variables should
be referenced by position.
}
\details{
By default, only top-level elements are numbered. Recursive numbering
of nested list elements can be enabled with \code{recursive = TRUE}.
}
\examples{
# Data frame
strX(mtcars)

# Nested list
x <- list(
  a = 1,
  b = list(
    c = 2,
    d = 3
  )
)

strX(x)

# Recursive numbering
strX(x, recursive = TRUE)

}
\seealso{
Other pkg.introspection:
\code{\link[=extractArgs]{extractArgs()}},
\code{\link[=funArgs]{funArgs()}},
\code{\link[=funCalls]{funCalls()}},
\code{\link[=funKeywords]{funKeywords()}},
\code{\link[=funList]{funList()}},
\code{\link[=getDotsArg]{getDotsArg()}},
\code{\link[=mergeArgs]{mergeArgs()}},
\code{\link[=rdLabels]{rdLabels()}},
\code{\link[=rdTitle]{rdTitle()}}
}
\concept{introspection}
\concept{pkg.introspection}
