% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strX.R
\name{strX}
\alias{strX}
\title{Compactly Display the Structure of any R Object}
\usage{
strX(x, ...)
}
\arguments{
\item{x}{any \code{R} object about which you want to have some information.}

\item{\dots}{dots are passed to \code{\link{str}}.}
}
\description{
Basically a wrapper for \code{\link{str}()}, extended with an enumeration
for the variables of a data.frame.
}
\examples{

strX(d.pizza)
}
\seealso{
\code{\link{str}}
}
\keyword{utilities}
