% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toBaseR.R
\name{toBaseR}
\alias{toBaseR}
\title{Back to Basics with Tibbles}
\usage{
toBaseR(x, ...)
}
\arguments{
\item{x}{the object to be converted.}

\item{\dots}{arguments passed on.}
}
\value{
converted object
}
\description{
Sometimes we might wish for the old days be back and want to work with
familiar objects. This function helps to convert \code{tibbles} to
\code{data.frames} as smoothly as possible.
}
\examples{

\dontrun{
# read a Stata file
url <- "http://www.stata.com/videos13/data/webclass.dta"
d.webclass <- toBaseR(haven::read_dta(url))

# read a SPSS file
url <- "https://stats.idre.ucla.edu/wp-content/uploads/2020/10/missing.sav"
d.miss <- toBaseR(haven::read_sav(url))
}
}
