% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toBaseR.R
\name{toBaseR}
\alias{toBaseR}
\title{Back to Basics with Tibbles}
\usage{
toBaseR(x, ...)
}
\arguments{
\item{x}{the object to be converted.}

\item{\dots}{arguments passed on.}
}
\value{
converted object
}
\description{
Sometimes we might wish for the old days be back and want to work with
familiar objects. This function helps to convert \code{tibbles} to
\code{data.frames} as smoothly as possible.
}
\examples{

\dontrun{
# read a Stata file
url <- "http://www.stata.com/videos13/data/webclass.dta"
d.webclass <- toBaseR(haven::read_dta(url))

# read a SPSS file
url <- "https://stats.idre.ucla.edu/wp-content/uploads/2020/10/missing.sav"
d.miss <- toBaseR(haven::read_sav(url))
}
}
\seealso{
Other data.manipulation: 
\code{\link{appendRowNames}()},
\code{\link{appendX}()},
\code{\link{as.array.xtabs}()},
\code{\link{columnWrap}()},
\code{\link{combLevels}()},
\code{\link{nf}()},
\code{\link{parseSASDatalines}()},
\code{\link{recodeX}()},
\code{\link{recycle}()},
\code{\link{revCode}()},
\code{\link{revX}()},
\code{\link{setNamesX}()},
\code{\link{set_strip_attr}},
\code{\link{sortX}()},
\code{\link{splitX}()},
\code{\link{stringsAsFactors}()},
\code{\link{tolong_towide}},
\code{\link{untable}()}
}
\concept{data-manipulation}
\concept{data-structures}
\concept{data.manipulation}
