% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-aliases.R
\name{num}
\alias{num}
\alias{int}
\alias{chr}
\alias{nchr}
\title{Coerce to numeric}
\usage{
num(x, ...)

int(x, ...)

chr(x, ...)

nchr(x)
}
\arguments{
\item{x}{a factor or character vector}

\item{...}{further arguments passed to \code{as.character}}
}
\description{
Shortcut for \code{as.numeric(as.character(x))}. Avoids returning
internal factor codes when coercing factors directly with \code{as.numeric}.
}
