% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unwhich.R
\name{unwhich}
\alias{unwhich}
\title{Inverse Which}
\usage{
unwhich(idx, n = max(idx), useNames = TRUE)
}
\arguments{
\item{idx}{the indices as returned by \code{\link{which}}.}

\item{n}{integer, the length of the original vector. This must not be less
than \code{max(idx)}, which is also the default.}

\item{useNames}{logical, determining if the names of the indices should be
preserved.}
}
\value{
a logical vector of the length n, with \code{TRUE} on the positions
\code{i}.
\item{comp1 }{Description of 'comp1'}
}
\description{
The inverse function to \code{\link{which}} creates a logical vector/matrix
from indices.
}
\examples{

ll <- c(TRUE, FALSE, TRUE, NA, FALSE, FALSE, TRUE)
names(ll) <- letters[seq(ll)]
i <- which(ll)
# back again (loosing the names of the FALSEs)
unwhich(i, length(ll))


}
\references{
\url{https://stackoverflow.com/questions/7659833/inverse-of-which}
}
\seealso{
\code{\link{which}}
}
\author{
Nick Sabbe
}
\keyword{attribute}
\keyword{logic}
