% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vRot.R
\name{vRot}
\alias{vRot}
\title{Rotate a vector}
\usage{
vRot(x, k = 1L)
}
\arguments{
\item{x}{A vector.}

\item{k}{Integer. Number of positions to rotate (default = 1).}
}
\value{
A vector of the same length as \code{x}, rotated cyclically.
}
\description{
Rotates a vector cyclically to the right by \code{k} positions.
Negative values of \code{k} rotate to the left.
}
\details{
The rotation is cyclic, meaning elements shifted off one end reappear on the other.
}
\examples{
vRot(1:5, 2)
# 4 5 1 2 3

vRot(1:5, -1)
# 2 3 4 5 1

}
\seealso{
Other vector.ops: 
\code{\link{Coalesce}()},
\code{\link{closest}()},
\code{\link{midx}()},
\code{\link{moveAvg}()},
\code{\link{naIf}()},
\code{\link{naReplace}()},
\code{\link{nz}()},
\code{\link{pairApply}()},
\code{\link{quot}()},
\code{\link{rankX}()},
\code{\link{splitAt}()},
\code{\link{trim}()},
\code{\link{unwhich}()},
\code{\link{vShift}()},
\code{\link{winsorize}()}
}
\concept{geometry}
\concept{mathematics}
\concept{vector-manipulation}
\concept{vector.ops}
