% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vShift.R
\name{vShift}
\alias{vShift}
\title{Shift a vector with NA padding}
\usage{
vShift(x, k = 1L)
}
\arguments{
\item{x}{A vector.}

\item{k}{Integer. Number of positions to shift.
Positive values shift to the right, negative values to the left.}
}
\value{
A vector of the same length as \code{x}, shifted with \code{NA} padding.
}
\description{
Shifts a vector to the left or right by \code{k} positions.
Vacated positions are filled with \code{NA}.
}
\details{
Unlike \code{VecRot()}, this function does not wrap elements around.
Elements shifted beyond the vector bounds are discarded.
}
\examples{
vShift(1:5, 2)
# NA NA 1 2 3

vShift(1:5, -2)
# 3 4 5 NA NA

vShift(1:5, 10)
# NA NA NA NA NA

}
\seealso{
Other vector.ops:
\code{\link[=Coalesce]{Coalesce()}},
\code{\link[=closest]{closest()}},
\code{\link[=midx]{midx()}},
\code{\link[=moveAvg]{moveAvg()}},
\code{\link[=naIf]{naIf()}},
\code{\link[=naReplace]{naReplace()}},
\code{\link[=nz]{nz()}},
\code{\link[=pairApply]{pairApply()}},
\code{\link[=quot]{quot()}},
\code{\link[=rankX]{rankX()}},
\code{\link[=splitAt]{splitAt()}},
\code{\link[=trim]{trim()}},
\code{\link[=unwhich]{unwhich()}},
\code{\link[=vRot]{vRot()}},
\code{\link[=winsorize]{winsorize()}}
}
\concept{time-series}
\concept{vector-manipulation}
\concept{vector.ops}
